+++ software/control/core/image_display.py
@@ -1,0 +1,225 @@
+# qt libraries
+from qtpy.QtCore import *
+from qtpy.QtWidgets import *
+from qtpy.QtGui import *
+
+from control._def import *
+
+from queue import Queue
+from threading import Thread, Lock
+import numpy as np
+import pyqtgraph as pg
+
+from typing import Optional, List, Union, Tuple
+
+class ImageDisplay(QObject):
+
+    image_to_display = Signal(np.ndarray)
+
+    def __init__(self):
+        QObject.__init__(self)
+        self.queue = Queue(10) # max 10 items in the queue
+        self.image_lock = Lock()
+        self.stop_signal_received = False
+        self.thread = Thread(target=self.process_queue) # type: ignore
+        self.thread.start()        
+        
+    def process_queue(self):
+        while True:
+            # stop the thread if stop signal is received
+            if self.stop_signal_received:
+                return
+            # process the queue
+            try:
+                [image,frame_ID,timestamp] = self.queue.get(timeout=0.1)
+                self.image_lock.acquire(True)
+                self.image_to_display.emit(image)
+                self.image_lock.release()
+                self.queue.task_done()
+            except:
+                pass
+
+    # def enqueue(self,image,frame_ID,timestamp):
+    def enqueue(self,image):
+        try:
+            self.queue.put_nowait([image,None,None])
+            # when using self.queue.put(str_) instead of try + nowait, program can be slowed down despite multithreading because of the block and the GIL
+            pass
+        except:
+            print('imageDisplay queue is full, image discarded')
+
+    def emit_directly(self,image):
+        self.image_to_display.emit(image)
+
+    def close(self):
+        self.queue.join()
+        self.stop_signal_received = True
+        self.thread.join()
+
+
+class ImageDisplayWindow(QMainWindow):
+
+    def __init__(self, window_title='', draw_crosshairs = False, show_LUT=False, autoLevels=False):
+        super().__init__()
+        self.setWindowTitle(window_title)
+        self.setWindowFlags(self.windowFlags() | Qt.CustomizeWindowHint)
+        self.setWindowFlags(self.windowFlags() & ~Qt.WindowCloseButtonHint)
+        self.widget = QWidget()
+        self.show_LUT = show_LUT
+        self.autoLevels = autoLevels
+
+        # interpret image data as row-major instead of col-major
+        pg.setConfigOptions(imageAxisOrder='row-major')
+
+        self.graphics_widget = pg.GraphicsLayoutWidget()
+        self.graphics_widget.view = self.graphics_widget.addViewBox()
+        self.graphics_widget.view.invertY()
+        
+        ## lock the aspect ratio so pixels are always square
+        self.graphics_widget.view.setAspectLocked(True)
+        
+        ## Create image item
+        if self.show_LUT:
+            self.graphics_widget.view = pg.ImageView()
+            self.graphics_widget.img = self.graphics_widget.view.getImageItem()
+            self.graphics_widget.img.setBorder('w')
+            self.graphics_widget.view.ui.roiBtn.hide()
+            self.graphics_widget.view.ui.menuBtn.hide()
+            # self.LUTWidget = self.graphics_widget.view.getHistogramWidget()
+            # self.LUTWidget.autoHistogramRange()
+            # self.graphics_widget.view.autolevels()
+        else:
+            self.graphics_widget.img = pg.ImageItem(border='w')
+            self.graphics_widget.view.addItem(self.graphics_widget.img)
+
+        ## Create ROI
+        self.roi_pos = (500,500)
+        self.roi_size = (500,500)
+        self.ROI = pg.ROI(self.roi_pos, self.roi_size, scaleSnap=True, translateSnap=True)
+        self.ROI.setZValue(10)
+        self.ROI.addScaleHandle((0,0), (1,1))
+        self.ROI.addScaleHandle((1,1), (0,0))
+        self.graphics_widget.view.addItem(self.ROI)
+        self.ROI.hide()
+        self.ROI.sigRegionChanged.connect(self.update_ROI)
+        self.roi_pos = self.ROI.pos()
+        self.roi_size = self.ROI.size()
+
+        ## Variables for annotating images
+        self.draw_rectangle = False
+        self.ptRect1 = None
+        self.ptRect2 = None
+        self.DrawCirc = False
+        self.centroid = None
+        self.DrawCrossHairs = False
+        self.image_offset = np.array([0, 0])
+
+        ## Layout
+        layout = QGridLayout()
+        if self.show_LUT:
+            layout.addWidget(self.graphics_widget.view, 0, 0) 
+        else:
+            layout.addWidget(self.graphics_widget, 0, 0) 
+        self.widget.setLayout(layout)
+        self.setCentralWidget(self.widget)
+
+        # set window size
+        desktopWidget = QDesktopWidget();
+        width = min(desktopWidget.height()*0.9,1000) #@@@TO MOVE@@@#
+        height = width
+        self.setFixedSize(width,height)
+
+    def display_image(self,image):
+        self.graphics_widget.img.setImage(image,autoLevels=self.autoLevels)
+
+    def update_ROI(self):
+        self.roi_pos = self.ROI.pos()
+        self.roi_size = self.ROI.size()
+
+    def show_ROI_selector(self):
+        self.ROI.show()
+
+    def hide_ROI_selector(self):
+        self.ROI.hide()
+
+    def get_roi(self):
+        return self.roi_pos,self.roi_size
+
+    def update_bounding_box(self,pts):
+        self.draw_rectangle=True
+        self.ptRect1=(pts[0][0],pts[0][1])
+        self.ptRect2=(pts[1][0],pts[1][1])
+
+    def get_roi_bounding_box(self):
+        self.update_ROI()
+        width = self.roi_size[0]
+        height = self.roi_size[1]
+        xmin = max(0, self.roi_pos[0])
+        ymin = max(0, self.roi_pos[1])
+        return np.array([xmin, ymin, width, height])
+
+    def set_autolevel(self,enabled):
+        self.autoLevels = enabled
+        print('set autolevel to ' + str(enabled))
+
+
+class ImageArrayDisplayWindow(QMainWindow):
+
+    def __init__(self, window_title=''):
+        super().__init__()
+        self.setWindowTitle(window_title)
+        self.setWindowFlags(self.windowFlags() | Qt.CustomizeWindowHint)
+        self.setWindowFlags(self.windowFlags() & ~Qt.WindowCloseButtonHint)
+        self.widget = QWidget()
+
+        # interpret image data as row-major instead of col-major
+        pg.setConfigOptions(imageAxisOrder='row-major')
+
+        self.graphics_widget_1 = pg.GraphicsLayoutWidget()
+        self.graphics_widget_1.view = self.graphics_widget_1.addViewBox()
+        self.graphics_widget_1.view.setAspectLocked(True)
+        self.graphics_widget_1.img = pg.ImageItem(border='w')
+        self.graphics_widget_1.view.addItem(self.graphics_widget_1.img)
+
+        self.graphics_widget_2 = pg.GraphicsLayoutWidget()
+        self.graphics_widget_2.view = self.graphics_widget_2.addViewBox()
+        self.graphics_widget_2.view.setAspectLocked(True)
+        self.graphics_widget_2.img = pg.ImageItem(border='w')
+        self.graphics_widget_2.view.addItem(self.graphics_widget_2.img)
+
+        self.graphics_widget_3 = pg.GraphicsLayoutWidget()
+        self.graphics_widget_3.view = self.graphics_widget_3.addViewBox()
+        self.graphics_widget_3.view.setAspectLocked(True)
+        self.graphics_widget_3.img = pg.ImageItem(border='w')
+        self.graphics_widget_3.view.addItem(self.graphics_widget_3.img)
+
+        self.graphics_widget_4 = pg.GraphicsLayoutWidget()
+        self.graphics_widget_4.view = self.graphics_widget_4.addViewBox()
+        self.graphics_widget_4.view.setAspectLocked(True)
+        self.graphics_widget_4.img = pg.ImageItem(border='w')
+        self.graphics_widget_4.view.addItem(self.graphics_widget_4.img)
+
+        ## Layout
+        layout = QGridLayout()
+        layout.addWidget(self.graphics_widget_1, 0, 0)
+        layout.addWidget(self.graphics_widget_2, 0, 1)
+        layout.addWidget(self.graphics_widget_3, 1, 0)
+        layout.addWidget(self.graphics_widget_4, 1, 1) 
+        self.widget.setLayout(layout)
+        self.setCentralWidget(self.widget)
+
+        # set window size
+        desktopWidget = QDesktopWidget()
+        width = min(desktopWidget.height()*0.9,1000) #@@@TO MOVE@@@#
+        height = width
+        self.setFixedSize(width,height)
+
+    def display_image(self,image,illumination_source):
+        if illumination_source < 11:
+            self.graphics_widget_1.img.setImage(image,autoLevels=False)
+        elif illumination_source == 11:
+            self.graphics_widget_2.img.setImage(image,autoLevels=False)
+        elif illumination_source == 12:
+            self.graphics_widget_3.img.setImage(image,autoLevels=False)
+        elif illumination_source == 13:
+            self.graphics_widget_4.img.setImage(image,autoLevels=False)
