+++ software/control/core/live_controller.py
@@ -1,0 +1,180 @@
+# qt libraries
+from qtpy.QtCore import *
+from qtpy.QtWidgets import *
+from qtpy.QtGui import *
+
+from control._def import *
+import time
+
+from typing import Optional, List, Union, Tuple
+
+import control.microcontroller as microcontroller
+import control.camera as camera
+from control.core import ConfigurationManager, Configuration
+
+class LiveController(QObject):
+
+    def __init__(self,
+        camera:camera.Camera,
+        microcontroller:microcontroller.Microcontroller,
+        configurationManager:ConfigurationManager,
+        control_illumination:bool=True,
+        use_internal_timer_for_hardware_trigger:bool=True
+    ):
+        QObject.__init__(self)
+        self.camera = camera
+        self.microcontroller = microcontroller
+        self.configurationManager:ConfigurationManager = configurationManager
+        self.currentConfiguration:Optional[Configuration] = None
+        self.trigger_mode = TriggerMode.SOFTWARE # @@@ change to None
+        self.is_live:bool = False
+        self.control_illumination = control_illumination
+        self.illumination_on:bool = False
+        self.use_internal_timer_for_hardware_trigger = use_internal_timer_for_hardware_trigger # use QTimer vs timer in the MCU
+
+        self.fps_trigger = 1
+        self.timer_trigger_interval = (1/self.fps_trigger)*1000
+
+        self.timer_trigger = QTimer()
+        self.timer_trigger.setInterval(self.timer_trigger_interval)
+        self.timer_trigger.timeout.connect(self.trigger_acquisition)
+
+        self.trigger_ID = -1
+
+        self.fps_real = 0
+        self.counter = 0
+        self.timestamp_last = 0
+
+        self.display_resolution_scaling = DEFAULT_DISPLAY_CROP/100
+
+    # illumination control
+    def turn_on_illumination(self):
+        self.microcontroller.turn_on_illumination()
+        self.illumination_on = True
+
+    def turn_off_illumination(self):
+        self.microcontroller.turn_off_illumination()
+        self.illumination_on = False
+
+    def set_illumination(self,illumination_source:int,intensity:int):
+        if illumination_source < 10: # LED matrix
+            self.microcontroller.set_illumination_led_matrix(illumination_source,r=(intensity/100)*LED_MATRIX_R_FACTOR,g=(intensity/100)*LED_MATRIX_G_FACTOR,b=(intensity/100)*LED_MATRIX_B_FACTOR)
+        else:
+            self.microcontroller.set_illumination(illumination_source,intensity)
+
+    def start_live(self):
+        self.is_live = True
+        self.camera.is_live = True
+        self.camera.start_streaming()
+        if self.trigger_mode == TriggerMode.SOFTWARE or ( self.trigger_mode == TriggerMode.HARDWARE and self.use_internal_timer_for_hardware_trigger ):
+            self._start_triggerred_acquisition()
+
+    def stop_live(self):
+        if self.is_live:
+            self.is_live = False
+            self.camera.is_live = False
+            if self.trigger_mode == TriggerMode.SOFTWARE:
+                self._stop_triggerred_acquisition()
+            # self.camera.stop_streaming() # 20210113 this line seems to cause problems when using af with multipoint
+            if self.trigger_mode == TriggerMode.CONTINUOUS:
+                self.camera.stop_streaming()
+            if ( self.trigger_mode == TriggerMode.SOFTWARE ) or ( self.trigger_mode == TriggerMode.HARDWARE and self.use_internal_timer_for_hardware_trigger ):
+                self._stop_triggerred_acquisition()
+            if self.control_illumination:
+                self.turn_off_illumination()
+
+    # software trigger related
+    def trigger_acquisition(self):
+        if self.trigger_mode == TriggerMode.SOFTWARE:
+            if self.control_illumination and self.illumination_on == False:
+                self.turn_on_illumination()
+            self.trigger_ID = self.trigger_ID + 1
+            self.camera.send_trigger()
+            # measure real fps
+            timestamp_now = round(time.time())
+            if timestamp_now == self.timestamp_last:
+                self.counter = self.counter+1
+            else:
+                self.timestamp_last = timestamp_now
+                self.fps_real = self.counter
+                self.counter = 0
+                # print('real trigger fps is ' + str(self.fps_real))
+        elif self.trigger_mode == TriggerMode.HARDWARE:
+            self.trigger_ID = self.trigger_ID + 1
+            self.microcontroller.send_hardware_trigger(control_illumination=True,illumination_on_time_us=self.camera.exposure_time*1000)
+
+    def _start_triggerred_acquisition(self):
+        self.timer_trigger.start()
+
+    def _set_trigger_fps(self,fps_trigger):
+        self.fps_trigger = fps_trigger
+        self.timer_trigger_interval = (1/self.fps_trigger)*1000
+        self.timer_trigger.setInterval(self.timer_trigger_interval)
+
+    def _stop_triggerred_acquisition(self):
+        self.timer_trigger.stop()
+
+    # trigger mode and settings
+    def set_trigger_mode(self,mode:TriggerMode):
+        if mode == TriggerMode.SOFTWARE:
+            if self.is_live and ( self.trigger_mode == TriggerMode.HARDWARE and self.use_internal_timer_for_hardware_trigger ):
+                self._stop_triggerred_acquisition()
+            self.camera.set_software_triggered_acquisition()
+            if self.is_live:
+                self._start_triggerred_acquisition()
+        if mode == TriggerMode.HARDWARE:
+            if self.trigger_mode == TriggerMode.SOFTWARE and self.is_live:
+                self._stop_triggerred_acquisition()
+            # self.camera.reset_camera_acquisition_counter()
+            self.camera.set_hardware_triggered_acquisition()
+            self.microcontroller.set_strobe_delay_us(self.camera.strobe_delay_us)
+            if self.is_live and self.use_internal_timer_for_hardware_trigger:
+                self._start_triggerred_acquisition()
+        if mode == TriggerMode.CONTINUOUS: 
+            if ( self.trigger_mode == TriggerMode.SOFTWARE ) or ( self.trigger_mode == TriggerMode.HARDWARE and self.use_internal_timer_for_hardware_trigger ):
+                self._stop_triggerred_acquisition()
+            self.camera.set_continuous_acquisition()
+        self.trigger_mode = mode
+
+    def set_trigger_fps(self,fps):
+        if ( self.trigger_mode == TriggerMode.SOFTWARE ) or ( self.trigger_mode == TriggerMode.HARDWARE and self.use_internal_timer_for_hardware_trigger ):
+            self._set_trigger_fps(fps)
+    
+    # set microscope mode
+    # @@@ to do: change softwareTriggerGenerator to TriggerGeneratror
+    def set_microscope_mode(self,configuration:Configuration):
+
+        self.currentConfiguration = configuration
+        print("setting microscope mode to " + self.currentConfiguration.name)
+        
+        # temporarily stop live while changing mode
+        if self.is_live is True:
+            self.timer_trigger.stop()
+            if self.control_illumination:
+                self.turn_off_illumination()
+
+        # set camera exposure time and analog gain
+        self.camera.set_exposure_time(self.currentConfiguration.exposure_time)
+        self.camera.set_analog_gain(self.currentConfiguration.analog_gain)
+
+        # set illumination
+        if self.control_illumination:
+            self.set_illumination(self.currentConfiguration.illumination_source,self.currentConfiguration.illumination_intensity)
+
+        # restart live 
+        if self.is_live is True:
+            if self.control_illumination:
+                self.turn_on_illumination()
+            self.timer_trigger.start()
+
+    def get_trigger_mode(self):
+        return self.trigger_mode
+
+    # slot
+    def on_new_frame(self):
+        if self.fps_trigger <= 5:
+            if self.control_illumination and self.illumination_on == True:
+                self.turn_off_illumination()
+
+    def set_display_resolution_scaling(self, display_resolution_scaling):
+        self.display_resolution_scaling = display_resolution_scaling/100
