+++ software/control/core/slide_position_controller.py
@@ -1,0 +1,171 @@
+# qt libraries
+from qtpy.QtCore import *
+from qtpy.QtWidgets import *
+from qtpy.QtGui import *
+
+import control.utils as utils
+from control._def import *
+
+import time
+
+from typing import Optional, List, Union, Tuple
+
+from control.core import NavigationController, LiveController
+
+class SlidePositionControlWorker(QObject):
+    
+    finished = Signal()
+    signal_stop_live = Signal()
+    signal_resume_live = Signal()
+
+    def __init__(self,slidePositionController,home_x_and_y_separately:bool=False):
+        QObject.__init__(self)
+        self.slidePositionController = slidePositionController
+        self.navigationController = slidePositionController.navigationController
+        self.microcontroller = self.navigationController.microcontroller
+        self.liveController = self.slidePositionController.liveController
+        self.home_x_and_y_separately = home_x_and_y_separately
+
+    def wait_till_operation_is_completed(self,timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S):
+        while self.microcontroller.is_busy():
+            time.sleep(SLEEP_TIME_S)
+            if time.time() - timestamp_start > SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S:
+                print('Error - slide position switching timeout, the program will exit')
+                self.navigationController.move_x(0)
+                self.navigationController.move_y(0)
+                exit()
+
+    def move_to_slide_loading_position(self):
+        was_live = self.liveController.is_live
+        if was_live:
+            self.signal_stop_live.emit()
+        if self.slidePositionController.homing_done == False or SLIDE_POTISION_SWITCHING_HOME_EVERYTIME:
+            if self.home_x_and_y_separately:
+                timestamp_start = time.time()
+                self.navigationController.home_x()
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+                self.navigationController.zero_x()
+                self.navigationController.move_x(SLIDE_POSITION.LOADING_X_MM)
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+                self.navigationController.home_y()
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+                self.navigationController.zero_y()
+                self.navigationController.move_y(SLIDE_POSITION.LOADING_Y_MM)
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+            else:
+                timestamp_start = time.time()
+                self.navigationController.home_xy()
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+                self.navigationController.zero_x()
+                self.navigationController.zero_y()
+                self.navigationController.move_x(SLIDE_POSITION.LOADING_X_MM)
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+                self.navigationController.move_y(SLIDE_POSITION.LOADING_Y_MM)
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+            self.slidePositionController.homing_done = True
+        else:
+            timestamp_start = time.time()
+            self.navigationController.move_y(SLIDE_POSITION.LOADING_Y_MM-self.navigationController.y_pos_mm)
+            self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+            self.navigationController.move_x(SLIDE_POSITION.LOADING_X_MM-self.navigationController.x_pos_mm)
+            self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+        if was_live:
+            self.signal_resume_live.emit()
+        self.slidePositionController.slide_loading_position_reached = True
+        self.finished.emit()
+
+    def move_to_slide_scanning_position(self):
+        was_live = self.liveController.is_live
+        if was_live:
+            self.signal_stop_live.emit()
+        if self.slidePositionController.homing_done == False or SLIDE_POTISION_SWITCHING_HOME_EVERYTIME:
+            if self.home_x_and_y_separately:
+                timestamp_start = time.time()
+                self.navigationController.home_y()
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+                self.navigationController.zero_y()
+                self.navigationController.move_y(SLIDE_POSITION.SCANNING_Y_MM)
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+                self.navigationController.home_x()
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+                self.navigationController.zero_x()
+                self.navigationController.move_x(SLIDE_POSITION.SCANNING_X_MM)
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+            else:
+                timestamp_start = time.time()
+                self.navigationController.home_xy()
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+                self.navigationController.zero_x()
+                self.navigationController.zero_y()
+                self.navigationController.move_y(SLIDE_POSITION.SCANNING_Y_MM)
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+                self.navigationController.move_x(SLIDE_POSITION.SCANNING_X_MM)
+                self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+            self.slidePositionController.homing_done = True
+        else:
+            timestamp_start = time.time()
+            self.navigationController.move_y(SLIDE_POSITION.SCANNING_Y_MM-self.navigationController.y_pos_mm)
+            self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+            self.navigationController.move_x(SLIDE_POSITION.SCANNING_X_MM-self.navigationController.x_pos_mm)
+            self.wait_till_operation_is_completed(timestamp_start, SLIDE_POTISION_SWITCHING_TIMEOUT_LIMIT_S)
+        if was_live:
+            self.signal_resume_live.emit()
+        self.slidePositionController.slide_scanning_position_reached = True
+        self.finished.emit()
+
+class SlidePositionController(QObject):
+
+    signal_slide_loading_position_reached = Signal()
+    signal_slide_scanning_position_reached = Signal()
+    signal_clear_slide = Signal()
+
+    def __init__(self,navigationController,liveController):
+        QObject.__init__(self)
+        self.navigationController:NavigationController = navigationController
+        self.liveController:LiveController = liveController
+        self.slide_loading_position_reached:bool = False
+        self.slide_scanning_position_reached:bool = False
+        self.homing_done:bool = False
+
+    def move_to_slide_loading_position(self):
+        # create a QThread object
+        self.thread = QThread()
+        # create a worker object
+        self.slidePositionControlWorker = SlidePositionControlWorker(self)
+        # move the worker to the thread
+        self.slidePositionControlWorker.moveToThread(self.thread)
+        # connect signals and slots
+        self.thread.started.connect(self.slidePositionControlWorker.move_to_slide_loading_position)
+        self.slidePositionControlWorker.signal_stop_live.connect(self.slot_stop_live,type=Qt.BlockingQueuedConnection)
+        self.slidePositionControlWorker.signal_resume_live.connect(self.slot_resume_live,type=Qt.BlockingQueuedConnection)
+        self.slidePositionControlWorker.finished.connect(self.signal_slide_loading_position_reached.emit)
+        self.slidePositionControlWorker.finished.connect(self.slidePositionControlWorker.deleteLater)
+        self.slidePositionControlWorker.finished.connect(self.thread.quit)
+        self.thread.finished.connect(self.thread.quit)
+        # start the thread
+        self.thread.start()
+
+    def move_to_slide_scanning_position(self):
+        # create a QThread object
+        self.thread = QThread()
+        # create a worker object
+        self.slidePositionControlWorker = SlidePositionControlWorker(self)
+        # move the worker to the thread
+        self.slidePositionControlWorker.moveToThread(self.thread)
+        # connect signals and slots
+        self.thread.started.connect(self.slidePositionControlWorker.move_to_slide_scanning_position)
+        self.slidePositionControlWorker.signal_stop_live.connect(self.slot_stop_live,type=Qt.BlockingQueuedConnection)
+        self.slidePositionControlWorker.signal_resume_live.connect(self.slot_resume_live,type=Qt.BlockingQueuedConnection)
+        self.slidePositionControlWorker.finished.connect(self.signal_slide_scanning_position_reached.emit)
+        self.slidePositionControlWorker.finished.connect(self.slidePositionControlWorker.deleteLater)
+        self.slidePositionControlWorker.finished.connect(self.thread.quit)
+        self.thread.finished.connect(self.thread.quit)
+        # start the thread
+        self.thread.start()
+        self.signal_clear_slide.emit()
+
+    def slot_stop_live(self):
+        self.liveController.stop_live()
+
+    def slot_resume_live(self):
+        self.liveController.start_live()
