+++ software/control/core/stream_handler.py
@@ -1,0 +1,146 @@
+# qt libraries
+from qtpy.QtCore import *
+from qtpy.QtWidgets import *
+from qtpy.QtGui import *
+
+import control.utils as utils
+from control._def import *
+
+from queue import Queue
+from threading import Thread, Lock
+import time
+import numpy as np
+import pyqtgraph as pg
+import cv2
+from datetime import datetime
+
+from lxml import etree as ET
+from pathlib import Path
+import control.utils_config as utils_config
+
+import math
+import json
+import pandas as pd
+
+import imageio as iio
+
+from typing import Optional, List, Union, Tuple
+import control.widgets as widgets
+
+import control.microcontroller as microcontroller
+import control.camera as camera
+
+class StreamHandler(QObject):
+
+    image_to_display = Signal(np.ndarray)
+    packet_image_to_write = Signal(np.ndarray, int, float)
+    packet_image_for_tracking = Signal(np.ndarray, int, float)
+    signal_new_frame_received = Signal()
+
+    def __init__(self,
+        crop_width:int=Acquisition.CROP_WIDTH,
+        crop_height:int=Acquisition.CROP_HEIGHT,
+        display_resolution_scaling:float=1.0
+    ):
+        QObject.__init__(self)
+        self.fps_display = 1
+        self.fps_save = 1
+        self.fps_track = 1
+        self.timestamp_last_display = 0
+        self.timestamp_last_save = 0
+        self.timestamp_last_track = 0
+
+        self.crop_width = crop_width
+        self.crop_height = crop_height
+        self.display_resolution_scaling = display_resolution_scaling
+
+        self.save_image_flag = False
+        self.track_flag = False
+        self.handler_busy = False
+
+        # for fps measurement
+        self.timestamp_last = 0
+        self.counter = 0
+        self.fps_real = 0
+
+    def start_recording(self):
+        self.save_image_flag = True
+
+    def stop_recording(self):
+        self.save_image_flag = False
+
+    def start_tracking(self):
+        self.tracking_flag = True
+
+    def stop_tracking(self):
+        self.tracking_flag = False
+
+    def set_display_fps(self,fps):
+        self.fps_display = fps
+
+    def set_save_fps(self,fps):
+        self.fps_save = fps
+
+    def set_crop(self,crop_width,crop_height):
+        self.crop_width = crop_width
+        self.crop_height = crop_height
+
+    def set_display_resolution_scaling(self, display_resolution_scaling):
+        self.display_resolution_scaling = display_resolution_scaling/100
+        print(self.display_resolution_scaling)
+
+    def on_new_frame(self, camera):
+
+        if camera.is_live:
+
+            camera.image_locked = True
+            self.handler_busy = True
+            self.signal_new_frame_received.emit() # self.liveController.turn_off_illumination()
+
+            # measure real fps
+            timestamp_now = round(time.time())
+            if timestamp_now == self.timestamp_last:
+                self.counter = self.counter+1
+            else:
+                self.timestamp_last = timestamp_now
+                self.fps_real = self.counter
+                self.counter = 0
+                # print('real camera fps is ' + str(self.fps_real))
+
+            # moved down (so that it does not modify the camera.current_frame, which causes minor problems for simulation) - 1/30/2022
+            # # rotate and flip - eventually these should be done in the camera
+            # camera.current_frame = utils.rotate_and_flip_image(camera.current_frame,rotate_image_angle=camera.rotate_image_angle,flip_image=camera.flip_image)
+
+            # crop image
+            image_cropped = utils.crop_image(camera.current_frame,self.crop_width,self.crop_height)
+            image_cropped = np.squeeze(image_cropped)
+            
+            # # rotate and flip - moved up (1/10/2022)
+            # image_cropped = utils.rotate_and_flip_image(image_cropped,rotate_image_angle=ROTATE_IMAGE_ANGLE,flip_image=FLIP_IMAGE)
+            # added on 1/30/2022
+            # @@@ to move to camera 
+            image_cropped = utils.rotate_and_flip_image(image_cropped,rotate_image_angle=camera.rotate_image_angle,flip_image=camera.flip_image)
+
+            # send image to display
+            time_now = time.time()
+            if time_now-self.timestamp_last_display >= 1/self.fps_display:
+                # self.image_to_display.emit(cv2.resize(image_cropped,(round(self.crop_width*self.display_resolution_scaling), round(self.crop_height*self.display_resolution_scaling)),cv2.INTER_LINEAR))
+                self.image_to_display.emit(utils.crop_image(image_cropped,round(self.crop_width*self.display_resolution_scaling), round(self.crop_height*self.display_resolution_scaling)))
+                self.timestamp_last_display = time_now
+
+            # send image to write
+            if self.save_image_flag and time_now-self.timestamp_last_save >= 1/self.fps_save:
+                if camera.is_color:
+                    image_cropped = cv2.cvtColor(image_cropped,cv2.COLOR_RGB2BGR)
+                self.packet_image_to_write.emit(image_cropped,camera.frame_ID,camera.timestamp)
+                self.timestamp_last_save = time_now
+
+            # send image to track
+            if self.track_flag and time_now-self.timestamp_last_track >= 1/self.fps_track:
+                # track is a blocking operation - it needs to be
+                # @@@ will cropping before emitting the signal lead to speedup?
+                self.packet_image_for_tracking.emit(image_cropped,camera.frame_ID,camera.timestamp)
+                self.timestamp_last_track = time_now
+
+            self.handler_busy = False
+            camera.image_locked = False
