--- software/control/widgets.py
@@ -1,1216 +1,0 @@
-# set QT_API environment variable
-import os 
-os.environ["QT_API"] = "pyqt5"
-
-# qt libraries
-from qtpy.QtCore import *
-from qtpy.QtWidgets import *
-from qtpy.QtGui import *
-
-import pyqtgraph as pg
-
-from datetime import datetime
-
-from control._def import *
-import control.core as core
-
-from typing import Optional, Union, List, Tuple
-
-class CameraSettingsWidget(QFrame):
-
-    signal_camera_set_temperature = Signal(float)
-
-    def __init__(self, camera, include_gain_exposure_time = True, include_camera_temperature_setting = False, main=None, *args, **kwargs):
-
-        super().__init__(*args, **kwargs)
-        self.camera = camera
-        self.add_components(include_gain_exposure_time,include_camera_temperature_setting)        
-        # set frame style
-        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
-
-    def add_components(self,include_gain_exposure_time,include_camera_temperature_setting):
-
-        # add buttons and input fields
-        self.entry_exposureTime = QDoubleSpinBox()
-        self.entry_exposureTime.setMinimum(self.camera.EXPOSURE_TIME_MS_MIN) 
-        self.entry_exposureTime.setMaximum(self.camera.EXPOSURE_TIME_MS_MAX) 
-        self.entry_exposureTime.setSingleStep(1)
-        self.entry_exposureTime.setValue(20)
-        self.camera.set_exposure_time(20)
-
-        self.entry_analogGain = QDoubleSpinBox()
-        self.entry_analogGain.setMinimum(self.camera.GAIN_MIN) 
-        self.entry_analogGain.setMaximum(self.camera.GAIN_MAX) 
-        self.entry_analogGain.setSingleStep(self.camera.GAIN_STEP)
-        self.entry_analogGain.setValue(0)
-        self.camera.set_analog_gain(0)
-
-        self.dropdown_pixelFormat = QComboBox()
-        self.dropdown_pixelFormat.addItems(['MONO8','MONO12','MONO14','MONO16','BAYER_RG8','BAYER_RG12'])
-        if self.camera.pixel_format is not None:
-            self.dropdown_pixelFormat.setCurrentText(self.camera.pixel_format)
-        # to do: load and save pixel format in configurations
-
-        self.entry_ROI_offset_x = QSpinBox()
-        self.entry_ROI_offset_x.setValue(CAMERA.ROI_OFFSET_X_DEFAULT)
-        self.entry_ROI_offset_x.setFixedWidth(40)
-        self.entry_ROI_offset_x.setMinimum(-1500)
-        self.entry_ROI_offset_x.setMaximum(1500)
-        self.entry_ROI_offset_x.setKeyboardTracking(False)
-        self.entry_ROI_offset_y = QSpinBox()
-        self.entry_ROI_offset_y.setValue(CAMERA.ROI_OFFSET_Y_DEFAULT)
-        self.entry_ROI_offset_y.setFixedWidth(40)
-        self.entry_ROI_offset_y.setMinimum(-1500)
-        self.entry_ROI_offset_y.setMaximum(1500)
-        self.entry_ROI_offset_y.setKeyboardTracking(False)
-        self.entry_ROI_width = QSpinBox()
-        self.entry_ROI_width.setMaximum(4000)
-        self.entry_ROI_width.setValue(CAMERA.ROI_WIDTH_DEFAULT)
-        self.entry_ROI_width.setFixedWidth(60)
-        self.entry_ROI_width.setKeyboardTracking(False)
-        self.entry_ROI_height = QSpinBox()
-        self.entry_ROI_height.setMaximum(3000)
-        self.entry_ROI_height.setValue(CAMERA.ROI_HEIGHT_DEFAULT)
-        self.entry_ROI_height.setFixedWidth(60)
-        self.entry_ROI_height.setKeyboardTracking(False)
-        self.entry_temperature = QDoubleSpinBox()
-        self.entry_temperature.setMaximum(25)
-        self.entry_temperature.setMinimum(-50)
-        self.entry_temperature.setDecimals(1)
-        self.label_temperature_measured = QLabel()
-        # self.label_temperature_measured.setNum(0)
-        self.label_temperature_measured.setFrameStyle(QFrame.Panel | QFrame.Sunken)
-
-        # connection
-        self.entry_exposureTime.valueChanged.connect(self.camera.set_exposure_time)
-        self.entry_analogGain.valueChanged.connect(self.camera.set_analog_gain)
-        self.dropdown_pixelFormat.currentTextChanged.connect(self.camera.set_pixel_format)
-        self.entry_ROI_offset_x.valueChanged.connect(self.set_ROI)
-        self.entry_ROI_offset_y.valueChanged.connect(self.set_ROI)
-        self.entry_ROI_height.valueChanged.connect(self.set_ROI)
-        self.entry_ROI_width.valueChanged.connect(self.set_ROI)
-        self.entry_temperature.valueChanged.connect(self.signal_camera_set_temperature.emit)
-
-        # layout
-        grid_ctrl = QGridLayout()
-        if include_gain_exposure_time:
-            grid_ctrl.addWidget(QLabel('Exposure Time (ms)'), 0,0)
-            grid_ctrl.addWidget(self.entry_exposureTime, 0,1)
-            grid_ctrl.addWidget(QLabel('Analog Gain'), 1,0)
-            grid_ctrl.addWidget(self.entry_analogGain, 1,1)
-        grid_ctrl.addWidget(QLabel('Pixel Format'), 2,0)
-        grid_ctrl.addWidget(self.dropdown_pixelFormat, 2,1)
-        if include_camera_temperature_setting:
-            grid_ctrl.addWidget(QLabel('Set Temperature (C)'),3,0)
-            grid_ctrl.addWidget(self.entry_temperature,3,1)
-            grid_ctrl.addWidget(QLabel('Actual Temperature (C)'),3,2)
-            grid_ctrl.addWidget(self.label_temperature_measured,3,3)
-
-        hbox1 = QHBoxLayout()
-        hbox1.addWidget(QLabel('ROI'))
-        hbox1.addStretch()
-        hbox1.addWidget(QLabel('height'))
-        hbox1.addWidget(self.entry_ROI_height)
-        hbox1.addWidget(QLabel('width'))
-        hbox1.addWidget(self.entry_ROI_width)
-        
-        hbox1.addWidget(QLabel('offset y'))
-        hbox1.addWidget(self.entry_ROI_offset_y)
-        hbox1.addWidget(QLabel('offset x'))
-        hbox1.addWidget(self.entry_ROI_offset_x)
-
-        self.grid = QGridLayout()
-        self.grid.addLayout(grid_ctrl,0,0)
-        self.grid.addLayout(hbox1,1,0)
-        self.setLayout(self.grid)
-
-    def set_exposure_time(self,exposure_time):
-        self.entry_exposureTime.setValue(exposure_time)
-
-    def set_analog_gain(self,analog_gain):
-        self.entry_analogGain.setValue(analog_gain)
-
-    def set_ROI(self):
-        self.camera.set_ROI(self.entry_ROI_offset_x.value(),self.entry_ROI_offset_y.value(),self.entry_ROI_width.value(),self.entry_ROI_height.value())
-
-    def update_measured_temperature(self,temperature):
-        self.label_temperature_measured.setNum(temperature)
-
-
-# 'Live' button text
-LIVE_BUTTON_IDLE_TEXT="Start Live"
-LIVE_BUTTON_RUNNING_TEXT="Stop Live"
-
-class LiveControlWidget(QFrame):
-    signal_newExposureTime = Signal(float)
-    signal_newAnalogGain = Signal(float)
-    signal_autoLevelSetting = Signal(bool)
-
-    def __init__(self, 
-        streamHandler, 
-        liveController:core.LiveController,
-        configurationManager:core.ConfigurationManager, 
-        show_trigger_options:bool=True, 
-        show_display_options:bool=True, 
-        show_autolevel:bool = False, 
-        autolevel:bool=False
-    ):
-        super().__init__()
-        self.liveController = liveController
-        self.streamHandler = streamHandler
-        self.configurationManager = configurationManager
-        self.fps_trigger:float = 10
-        self.fps_display:float = 10
-        self.liveController.set_trigger_fps(self.fps_trigger)
-        self.streamHandler.set_display_fps(self.fps_display)
-        
-        self.triggerMode = TriggerMode.SOFTWARE
-        # note that this references the object in self.configurationManager.configurations
-        self.currentConfiguration = self.configurationManager.configurations[0]
-
-        self.add_components(show_trigger_options,show_display_options,show_autolevel,autolevel)
-        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
-        self.update_microscope_mode_by_name(self.currentConfiguration.name)
-
-        self.is_switching_mode = False # flag used to prevent from settings being set by twice - from both mode change slot and value change slot; another way is to use blockSignals(True)
-
-    def add_components(self,
-        show_trigger_options:bool,
-        show_display_options:bool,
-        show_autolevel:bool,
-        autolevel:bool
-    ):
-        # line 0: trigger mode
-        self.triggerMode = None
-        self.dropdown_triggerManu = QComboBox()
-        self.dropdown_triggerManu.addItems([TriggerMode.SOFTWARE,TriggerMode.HARDWARE,TriggerMode.CONTINUOUS])
-
-        # line 1: fps
-        self.entry_triggerFPS = QDoubleSpinBox()
-        self.entry_triggerFPS.setMinimum(0.02) 
-        self.entry_triggerFPS.setMaximum(1000) 
-        self.entry_triggerFPS.setSingleStep(1)
-        self.entry_triggerFPS.setValue(self.fps_trigger)
-
-        # line 2: choose microscope mode / toggle live mode 
-        self.dropdown_modeSelection = QComboBox()
-        for microscope_configuration in self.configurationManager.configurations:
-            self.dropdown_modeSelection.addItems([microscope_configuration.name])
-        self.dropdown_modeSelection.setCurrentText(self.currentConfiguration.name)
-
-        self.btn_live = QPushButton(LIVE_BUTTON_IDLE_TEXT)
-        self.btn_live.setCheckable(True)
-        self.btn_live.setChecked(False)
-        self.btn_live.setDefault(False)
-
-        # line 3: exposure time and analog gain associated with the current mode
-        self.entry_exposureTime = QDoubleSpinBox()
-        self.entry_exposureTime.setMinimum(self.liveController.camera.EXPOSURE_TIME_MS_MIN) 
-        self.entry_exposureTime.setMaximum(self.liveController.camera.EXPOSURE_TIME_MS_MAX) 
-        self.entry_exposureTime.setSingleStep(1)
-        self.entry_exposureTime.setValue(0)
-
-        self.entry_analogGain = QDoubleSpinBox()
-        self.entry_analogGain = QDoubleSpinBox()
-        self.entry_analogGain.setMinimum(0) 
-        self.entry_analogGain.setMaximum(24) 
-        self.entry_analogGain.setSingleStep(0.1)
-        self.entry_analogGain.setValue(0)
-
-        self.slider_illuminationIntensity = QSlider(Qt.Horizontal)
-        self.slider_illuminationIntensity.setTickPosition(QSlider.TicksBelow)
-        self.slider_illuminationIntensity.setMinimum(0)
-        self.slider_illuminationIntensity.setMaximum(100)
-        self.slider_illuminationIntensity.setValue(100)
-        self.slider_illuminationIntensity.setSingleStep(0.1)
-
-        self.entry_illuminationIntensity = QDoubleSpinBox()
-        self.entry_illuminationIntensity.setMinimum(0.1) 
-        self.entry_illuminationIntensity.setMaximum(100) 
-        self.entry_illuminationIntensity.setSingleStep(0.1)
-        self.entry_illuminationIntensity.setValue(100)
-
-        # line 4: display fps and resolution scaling
-        self.entry_displayFPS = QDoubleSpinBox()
-        self.entry_displayFPS.setMinimum(1) 
-        self.entry_displayFPS.setMaximum(240) 
-        self.entry_displayFPS.setSingleStep(1)
-        self.entry_displayFPS.setValue(self.fps_display)
-
-        self.slider_resolutionScaling = QSlider(Qt.Horizontal)
-        self.slider_resolutionScaling.setTickPosition(QSlider.TicksBelow)
-        self.slider_resolutionScaling.setMinimum(10)
-        self.slider_resolutionScaling.setMaximum(100)
-        self.slider_resolutionScaling.setValue(DEFAULT_DISPLAY_CROP)
-        self.slider_resolutionScaling.setSingleStep(10)
-
-        # autolevel
-        self.btn_autolevel = QPushButton('Autolevel')
-        self.btn_autolevel.setCheckable(True)
-        self.btn_autolevel.setChecked(autolevel)
-        
-        # connections
-        self.entry_triggerFPS.valueChanged.connect(self.liveController.set_trigger_fps)
-        self.entry_displayFPS.valueChanged.connect(self.streamHandler.set_display_fps)
-        self.slider_resolutionScaling.valueChanged.connect(self.streamHandler.set_display_resolution_scaling)
-        self.slider_resolutionScaling.valueChanged.connect(self.liveController.set_display_resolution_scaling)
-        self.dropdown_modeSelection.currentTextChanged.connect(self.update_microscope_mode_by_name)
-        self.dropdown_triggerManu.currentIndexChanged.connect(self.update_trigger_mode)
-        self.btn_live.clicked.connect(self.toggle_live)
-        self.entry_exposureTime.valueChanged.connect(self.update_config_exposure_time)
-        self.entry_analogGain.valueChanged.connect(self.update_config_analog_gain)
-        self.entry_illuminationIntensity.valueChanged.connect(self.update_config_illumination_intensity)
-        self.entry_illuminationIntensity.valueChanged.connect(self.slider_illuminationIntensity.setValue)
-        self.slider_illuminationIntensity.valueChanged.connect(self.entry_illuminationIntensity.setValue)
-        self.btn_autolevel.clicked.connect(self.signal_autoLevelSetting.emit)
-
-        # layout
-        grid_line0 = QGridLayout()
-        grid_line0.addWidget(QLabel('Trigger Mode'), 0,0)
-        grid_line0.addWidget(self.dropdown_triggerManu, 0,1)
-        grid_line0.addWidget(QLabel('Trigger FPS'), 0,2)
-        grid_line0.addWidget(self.entry_triggerFPS, 0,3)
-
-        grid_line1 = QGridLayout()
-        grid_line1.addWidget(QLabel('Imaging Mode'), 0,0)
-        grid_line1.addWidget(self.dropdown_modeSelection, 0,1)
-        grid_line1.addWidget(self.btn_live, 0,2)
-
-        grid_line2 = QGridLayout()
-        grid_line2.addWidget(QLabel('Exposure Time (ms)'), 0,0)
-        grid_line2.addWidget(self.entry_exposureTime, 0,1)
-        grid_line2.addWidget(QLabel('Analog Gain'), 0,2)
-        grid_line2.addWidget(self.entry_analogGain, 0,3)
-
-        grid_line4 = QGridLayout()
-        grid_line4.addWidget(QLabel('Illumination'), 0,0)
-        grid_line4.addWidget(self.slider_illuminationIntensity, 0,1)
-        grid_line4.addWidget(self.entry_illuminationIntensity, 0,2)
-
-        grid_line3 = QGridLayout()
-        grid_line3.addWidget(QLabel('Display FPS'), 0,0)
-        grid_line3.addWidget(self.entry_displayFPS, 0,1)
-        grid_line3.addWidget(QLabel('Display Resolution'), 0,2)
-        grid_line3.addWidget(self.slider_resolutionScaling,0,3)
-        if show_autolevel:
-            grid_line3.addWidget(self.btn_autolevel,0,4)
-
-        self.grid = QVBoxLayout()
-        if show_trigger_options:
-            self.grid.addLayout(grid_line0)
-        self.grid.addLayout(grid_line1)
-        self.grid.addLayout(grid_line2)
-        self.grid.addLayout(grid_line4)
-        if show_display_options:
-            self.grid.addLayout(grid_line3)
-        self.grid.addStretch()
-        self.setLayout(self.grid)
-
-    def toggle_live(self,pressed:bool):
-        if pressed:
-            self.btn_live.setText(LIVE_BUTTON_RUNNING_TEXT)
-            self.liveController.start_live()
-        else:
-            self.btn_live.setText(LIVE_BUTTON_IDLE_TEXT)
-            self.liveController.stop_live()
-
-    def update_camera_settings(self):
-        self.signal_newAnalogGain.emit(self.entry_analogGain.value())
-        self.signal_newExposureTime.emit(self.entry_exposureTime.value())
-
-    def update_microscope_mode_by_name(self,current_microscope_mode_name:str):
-        self.is_switching_mode = True
-        # identify the mode selected (note that this references the object in self.configurationManager.configurations)
-        self.currentConfiguration = next((config for config in self.configurationManager.configurations if config.name == current_microscope_mode_name), None)
-        # update the microscope to the current configuration
-        self.liveController.set_microscope_mode(self.currentConfiguration)
-        # update the exposure time and analog gain settings according to the selected configuration
-        self.entry_exposureTime.setValue(self.currentConfiguration.exposure_time)
-        self.entry_analogGain.setValue(self.currentConfiguration.analog_gain)
-        self.entry_illuminationIntensity.setValue(self.currentConfiguration.illumination_intensity)
-        self.is_switching_mode = False
-
-    def update_trigger_mode(self):
-        self.liveController.set_trigger_mode(self.dropdown_triggerManu.currentText())
-
-    def update_config_exposure_time(self,new_value:float):
-        if self.is_switching_mode == False:
-            self.currentConfiguration.exposure_time = new_value
-            self.configurationManager.update_configuration(self.currentConfiguration.id,'ExposureTime',new_value)
-            self.signal_newExposureTime.emit(new_value)
-
-    def update_config_analog_gain(self,new_value):
-        if self.is_switching_mode == False:
-            self.currentConfiguration.analog_gain = new_value
-            self.configurationManager.update_configuration(self.currentConfiguration.id,'AnalogGain',new_value)
-            self.signal_newAnalogGain.emit(new_value)
-
-    def update_config_illumination_intensity(self,new_value):
-        if self.is_switching_mode == False:
-            self.currentConfiguration.illumination_intensity = new_value
-            self.configurationManager.update_configuration(self.currentConfiguration.id,'IlluminationIntensity',new_value)
-            self.liveController.set_illumination(self.currentConfiguration.illumination_source, self.currentConfiguration.illumination_intensity)
-
-    def set_microscope_mode(self,config):
-        # self.liveController.set_microscope_mode(config)
-        self.dropdown_modeSelection.setCurrentText(config.name)
-
-    def set_trigger_mode(self,trigger_mode):
-        self.dropdown_triggerManu.setCurrentText(trigger_mode)
-        self.liveController.set_trigger_mode(self.dropdown_triggerManu.currentText())
-
-class RecordingWidget(QFrame):
-    def __init__(self, streamHandler, imageSaver, main=None, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.imageSaver = imageSaver # for saving path control
-        self.streamHandler = streamHandler
-        self.base_path_is_set = False
-        self.add_components()
-        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
-
-    def add_components(self):
-        self.btn_setSavingDir = QPushButton('Browse')
-        self.btn_setSavingDir.setDefault(False)
-        self.btn_setSavingDir.setIcon(QIcon('icon/folder.png'))
-        
-        self.lineEdit_savingDir = QLineEdit()
-        self.lineEdit_savingDir.setReadOnly(True)
-        self.lineEdit_savingDir.setText('Choose a base saving directory')
-
-        self.lineEdit_savingDir.setText(DEFAULT_SAVING_PATH)
-        self.imageSaver.set_base_path(DEFAULT_SAVING_PATH)
-
-        self.lineEdit_experimentID = QLineEdit()
-
-        self.entry_saveFPS = QDoubleSpinBox()
-        self.entry_saveFPS.setMinimum(0.02) 
-        self.entry_saveFPS.setMaximum(1000) 
-        self.entry_saveFPS.setSingleStep(1)
-        self.entry_saveFPS.setValue(1)
-        self.streamHandler.set_save_fps(1)
-
-        self.entry_timeLimit = QSpinBox()
-        self.entry_timeLimit.setMinimum(-1) 
-        self.entry_timeLimit.setMaximum(60*60*24*30) 
-        self.entry_timeLimit.setSingleStep(1)
-        self.entry_timeLimit.setValue(-1)
-
-        self.btn_record = QPushButton("Record")
-        self.btn_record.setCheckable(True)
-        self.btn_record.setChecked(False)
-        self.btn_record.setDefault(False)
-
-        grid_line1 = QGridLayout()
-        grid_line1.addWidget(QLabel('Saving Path'))
-        grid_line1.addWidget(self.lineEdit_savingDir, 0,1)
-        grid_line1.addWidget(self.btn_setSavingDir, 0,2)
-
-        grid_line2 = QGridLayout()
-        grid_line2.addWidget(QLabel('Experiment ID'), 0,0)
-        grid_line2.addWidget(self.lineEdit_experimentID,0,1)
-
-        grid_line3 = QGridLayout()
-        grid_line3.addWidget(QLabel('Saving FPS'), 0,0)
-        grid_line3.addWidget(self.entry_saveFPS, 0,1)
-        grid_line3.addWidget(QLabel('Time Limit (s)'), 0,2)
-        grid_line3.addWidget(self.entry_timeLimit, 0,3)
-        grid_line3.addWidget(self.btn_record, 0,4)
-
-        self.grid = QGridLayout()
-        self.grid.addLayout(grid_line1,0,0)
-        self.grid.addLayout(grid_line2,1,0)
-        self.grid.addLayout(grid_line3,2,0)
-        self.setLayout(self.grid)
-
-        # add and display a timer - to be implemented
-        # self.timer = QTimer()
-
-        # connections
-        self.btn_setSavingDir.clicked.connect(self.set_saving_dir)
-        self.btn_record.clicked.connect(self.toggle_recording)
-        self.entry_saveFPS.valueChanged.connect(self.streamHandler.set_save_fps)
-        self.entry_timeLimit.valueChanged.connect(self.imageSaver.set_recording_time_limit)
-        self.imageSaver.stop_recording.connect(self.stop_recording)
-
-    def set_saving_dir(self):
-        dialog = QFileDialog()
-        save_dir_base = dialog.getExistingDirectory(None, "Select Folder")
-        self.imageSaver.set_base_path(save_dir_base)
-        self.lineEdit_savingDir.setText(save_dir_base)
-        self.base_path_is_set = True
-
-    def toggle_recording(self,pressed):
-        if self.base_path_is_set == False:
-            self.btn_record.setChecked(False)
-            msg = QMessageBox()
-            msg.setText("Please choose base saving directory first")
-            msg.exec_()
-            return
-        if pressed:
-            self.lineEdit_experimentID.setEnabled(False)
-            self.btn_setSavingDir.setEnabled(False)
-            self.imageSaver.start_new_experiment(self.lineEdit_experimentID.text())
-            self.streamHandler.start_recording()
-        else:
-            self.streamHandler.stop_recording()
-            self.lineEdit_experimentID.setEnabled(True)
-            self.btn_setSavingDir.setEnabled(True)
-
-    # stop_recording can be called by imageSaver
-    def stop_recording(self):
-        self.lineEdit_experimentID.setEnabled(True)
-        self.btn_record.setChecked(False)
-        self.streamHandler.stop_recording()
-        self.btn_setSavingDir.setEnabled(True)
-
-class NavigationWidget(QFrame):
-    def __init__(self, navigationController, slidePositionController=None, main=None, widget_configuration = 'full', *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.navigationController = navigationController
-        self.slidePositionController = slidePositionController
-        self.widget_configuration = widget_configuration
-        self.slide_position = None
-        self.add_components()
-        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
-
-    def add_components(self):
-        self.label_Xpos = QLabel()
-        self.label_Xpos.setNum(0)
-        self.label_Xpos.setFrameStyle(QFrame.Panel | QFrame.Sunken)
-        self.entry_dX = QDoubleSpinBox()
-        self.entry_dX.setMinimum(0) 
-        self.entry_dX.setMaximum(25) 
-        self.entry_dX.setSingleStep(0.2)
-        self.entry_dX.setValue(0)
-        self.entry_dX.setDecimals(3)
-        self.entry_dX.setKeyboardTracking(False)
-        self.btn_moveX_forward = QPushButton('Forward')
-        self.btn_moveX_forward.setDefault(False)
-        self.btn_moveX_backward = QPushButton('Backward')
-        self.btn_moveX_backward.setDefault(False)
-
-        self.btn_home_X = QPushButton('Home X')
-        self.btn_home_X.setDefault(False)
-        self.btn_home_X.setEnabled(HOMING_ENABLED_X)
-        self.btn_zero_X = QPushButton('Zero X')
-        self.btn_zero_X.setDefault(False)
-        
-        self.label_Ypos = QLabel()
-        self.label_Ypos.setNum(0)
-        self.label_Ypos.setFrameStyle(QFrame.Panel | QFrame.Sunken)
-        self.entry_dY = QDoubleSpinBox()
-        self.entry_dY.setMinimum(0)
-        self.entry_dY.setMaximum(25)
-        self.entry_dY.setSingleStep(0.2)
-        self.entry_dY.setValue(0)
-        self.entry_dY.setDecimals(3)
-        self.entry_dY.setKeyboardTracking(False)
-        self.btn_moveY_forward = QPushButton('Forward')
-        self.btn_moveY_forward.setDefault(False)
-        self.btn_moveY_backward = QPushButton('Backward')
-        self.btn_moveY_backward.setDefault(False)
-
-        self.btn_home_Y = QPushButton('Home Y')
-        self.btn_home_Y.setDefault(False)
-        self.btn_home_Y.setEnabled(HOMING_ENABLED_Y)
-        self.btn_zero_Y = QPushButton('Zero Y')
-        self.btn_zero_Y.setDefault(False)
-
-        self.label_Zpos = QLabel()
-        self.label_Zpos.setNum(0)
-        self.label_Zpos.setFrameStyle(QFrame.Panel | QFrame.Sunken)
-        self.entry_dZ = QDoubleSpinBox()
-        self.entry_dZ.setMinimum(0) 
-        self.entry_dZ.setMaximum(1000) 
-        self.entry_dZ.setSingleStep(0.2)
-        self.entry_dZ.setValue(0)
-        self.entry_dZ.setDecimals(3)
-        self.entry_dZ.setKeyboardTracking(False)
-        self.btn_moveZ_forward = QPushButton('Forward')
-        self.btn_moveZ_forward.setDefault(False)
-        self.btn_moveZ_backward = QPushButton('Backward')
-        self.btn_moveZ_backward.setDefault(False)
-
-        self.btn_home_Z = QPushButton('Home Z')
-        self.btn_home_Z.setDefault(False)
-        self.btn_home_Z.setEnabled(HOMING_ENABLED_Z)
-        self.btn_zero_Z = QPushButton('Zero Z')
-        self.btn_zero_Z.setDefault(False)
-
-        self.btn_load_slide = QPushButton('To Slide Loading Position')
-        
-        grid_line0 = QGridLayout()
-        grid_line0.addWidget(QLabel('X (mm)'), 0,0)
-        grid_line0.addWidget(self.label_Xpos, 0,1)
-        grid_line0.addWidget(self.entry_dX, 0,2)
-        grid_line0.addWidget(self.btn_moveX_forward, 0,3)
-        grid_line0.addWidget(self.btn_moveX_backward, 0,4)
-        
-        grid_line1 = QGridLayout()
-        grid_line1.addWidget(QLabel('Y (mm)'), 0,0)
-        grid_line1.addWidget(self.label_Ypos, 0,1)
-        grid_line1.addWidget(self.entry_dY, 0,2)
-        grid_line1.addWidget(self.btn_moveY_forward, 0,3)
-        grid_line1.addWidget(self.btn_moveY_backward, 0,4)
-
-        grid_line2 = QGridLayout()
-        grid_line2.addWidget(QLabel('Z (um)'), 0,0)
-        grid_line2.addWidget(self.label_Zpos, 0,1)
-        grid_line2.addWidget(self.entry_dZ, 0,2)
-        grid_line2.addWidget(self.btn_moveZ_forward, 0,3)
-        grid_line2.addWidget(self.btn_moveZ_backward, 0,4)
-        
-        grid_line3 = QGridLayout()
-        if self.widget_configuration == 'full':
-            grid_line3.addWidget(self.btn_zero_X, 0,3)
-            grid_line3.addWidget(self.btn_zero_Y, 0,4)
-            grid_line3.addWidget(self.btn_zero_Z, 0,5)
-            grid_line3.addWidget(self.btn_home_X, 0,0)
-            grid_line3.addWidget(self.btn_home_Y, 0,1)
-            grid_line3.addWidget(self.btn_home_Z, 0,2)
-        elif self.widget_configuration == 'malaria':
-            grid_line3.addWidget(self.btn_load_slide, 0,0,1,2)
-            grid_line3.addWidget(self.btn_home_Z, 0,2,1,1)
-            grid_line3.addWidget(self.btn_zero_Z, 0,3,1,1)
-        elif self.widget_configuration == '384 well plate':
-            grid_line3.addWidget(self.btn_home_Z, 0,2,1,1)
-            grid_line3.addWidget(self.btn_zero_Z, 0,3,1,1)
-        elif self.widget_configuration == '96 well plate':
-            grid_line3.addWidget(self.btn_home_Z, 0,2,1,1)
-            grid_line3.addWidget(self.btn_zero_Z, 0,3,1,1)
-
-        self.grid = QGridLayout()
-        self.grid.addLayout(grid_line0,0,0)
-        self.grid.addLayout(grid_line1,1,0)
-        self.grid.addLayout(grid_line2,2,0)
-        self.grid.addLayout(grid_line3,3,0)
-        self.setLayout(self.grid)
-
-        self.entry_dX.valueChanged.connect(self.set_deltaX)
-        self.entry_dY.valueChanged.connect(self.set_deltaY)
-        self.entry_dZ.valueChanged.connect(self.set_deltaZ)
-
-        self.btn_moveX_forward.clicked.connect(self.move_x_forward)
-        self.btn_moveX_backward.clicked.connect(self.move_x_backward)
-        self.btn_moveY_forward.clicked.connect(self.move_y_forward)
-        self.btn_moveY_backward.clicked.connect(self.move_y_backward)
-        self.btn_moveZ_forward.clicked.connect(self.move_z_forward)
-        self.btn_moveZ_backward.clicked.connect(self.move_z_backward)
-
-        self.btn_home_X.clicked.connect(self.home_x)
-        self.btn_home_Y.clicked.connect(self.home_y)
-        self.btn_home_Z.clicked.connect(self.home_z)
-        self.btn_zero_X.clicked.connect(self.zero_x)
-        self.btn_zero_Y.clicked.connect(self.zero_y)
-        self.btn_zero_Z.clicked.connect(self.zero_z)
-
-        self.btn_load_slide.clicked.connect(self.switch_position)
-        self.btn_load_slide.setStyleSheet("background-color: #C2C2FF");
-        
-    def move_x_forward(self):
-        self.navigationController.move_x(self.entry_dX.value())
-    def move_x_backward(self):
-        self.navigationController.move_x(-self.entry_dX.value())
-    def move_y_forward(self):
-        self.navigationController.move_y(self.entry_dY.value())
-    def move_y_backward(self):
-        self.navigationController.move_y(-self.entry_dY.value())
-    def move_z_forward(self):
-        self.navigationController.move_z(self.entry_dZ.value()/1000)
-    def move_z_backward(self):
-        self.navigationController.move_z(-self.entry_dZ.value()/1000) 
-
-    def set_deltaX(self,value):
-        mm_per_ustep = SCREW_PITCH_X_MM/(self.navigationController.x_microstepping*FULLSTEPS_PER_REV_X) # to implement a get_x_microstepping() in multipointController
-        deltaX = round(value/mm_per_ustep)*mm_per_ustep
-        self.entry_dX.setValue(deltaX)
-    def set_deltaY(self,value):
-        mm_per_ustep = SCREW_PITCH_Y_MM/(self.navigationController.y_microstepping*FULLSTEPS_PER_REV_Y)
-        deltaY = round(value/mm_per_ustep)*mm_per_ustep
-        self.entry_dY.setValue(deltaY)
-    def set_deltaZ(self,value):
-        mm_per_ustep = SCREW_PITCH_Z_MM/(self.navigationController.z_microstepping*FULLSTEPS_PER_REV_Z)
-        deltaZ = round(value/1000/mm_per_ustep)*mm_per_ustep*1000
-        self.entry_dZ.setValue(deltaZ)
-
-    def home_x(self):
-        msg = QMessageBox()
-        msg.setIcon(QMessageBox.Information)
-        msg.setText("Confirm your action")
-        msg.setInformativeText("Click OK to run homing")
-        msg.setWindowTitle("Confirmation")
-        msg.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
-        msg.setDefaultButton(QMessageBox.Cancel)
-        retval = msg.exec_()
-        if QMessageBox.Ok == retval:
-            self.navigationController.home_x()
-
-    def home_y(self):
-        msg = QMessageBox()
-        msg.setIcon(QMessageBox.Information)
-        msg.setText("Confirm your action")
-        msg.setInformativeText("Click OK to run homing")
-        msg.setWindowTitle("Confirmation")
-        msg.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
-        msg.setDefaultButton(QMessageBox.Cancel)
-        retval = msg.exec_()
-        if QMessageBox.Ok == retval:
-            self.navigationController.home_y()
-
-    def home_z(self):
-        msg = QMessageBox()
-        msg.setIcon(QMessageBox.Information)
-        msg.setText("Confirm your action")
-        msg.setInformativeText("Click OK to run homing")
-        msg.setWindowTitle("Confirmation")
-        msg.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
-        msg.setDefaultButton(QMessageBox.Cancel)
-        retval = msg.exec_()
-        if QMessageBox.Ok == retval:
-            self.navigationController.home_z()
-
-    def zero_x(self):
-        self.navigationController.zero_x()
-
-    def zero_y(self):
-        self.navigationController.zero_y()
-
-    def zero_z(self):
-        self.navigationController.zero_z()
-
-    def slot_slide_loading_position_reached(self):
-        self.slide_position = 'loading'
-        self.btn_load_slide.setStyleSheet("background-color: #C2FFC2");
-        self.btn_load_slide.setText('To Slide Scanning Position')
-        self.btn_moveX_forward.setEnabled(False)
-        self.btn_moveX_backward.setEnabled(False)
-        self.btn_moveY_forward.setEnabled(False)
-        self.btn_moveY_backward.setEnabled(False)
-        self.btn_moveZ_forward.setEnabled(False)
-        self.btn_moveZ_backward.setEnabled(False)
-
-    def slot_slide_scanning_position_reached(self):
-        self.slide_position = 'scanning'
-        self.btn_load_slide.setStyleSheet("background-color: #C2C2FF");
-        self.btn_load_slide.setText('To Slide Loading Position')
-        self.btn_moveX_forward.setEnabled(True)
-        self.btn_moveX_backward.setEnabled(True)
-        self.btn_moveY_forward.setEnabled(True)
-        self.btn_moveY_backward.setEnabled(True)
-        self.btn_moveZ_forward.setEnabled(True)
-        self.btn_moveZ_backward.setEnabled(True)
-
-    def switch_position(self):
-        if self.slide_position != 'loading':
-            self.slidePositionController.move_to_slide_loading_position()
-        else:
-            self.slidePositionController.move_to_slide_scanning_position()
-
-class DACControWidget(QFrame):
-    def __init__(self, microcontroller ,*args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.microcontroller = microcontroller
-        self.add_components()
-        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
-
-    def add_components(self):
-        self.slider_DAC0 = QSlider(Qt.Horizontal)
-        self.slider_DAC0.setTickPosition(QSlider.TicksBelow)
-        self.slider_DAC0.setMinimum(0)
-        self.slider_DAC0.setMaximum(100)
-        self.slider_DAC0.setSingleStep(0.1)
-        self.slider_DAC0.setValue(0)
-
-        self.entry_DAC0 = QDoubleSpinBox()
-        self.entry_DAC0.setMinimum(0) 
-        self.entry_DAC0.setMaximum(100) 
-        self.entry_DAC0.setSingleStep(0.1)
-        self.entry_DAC0.setValue(0)
-        self.entry_DAC0.setKeyboardTracking(False)
-
-        self.slider_DAC1 = QSlider(Qt.Horizontal)
-        self.slider_DAC1.setTickPosition(QSlider.TicksBelow)
-        self.slider_DAC1.setMinimum(0)
-        self.slider_DAC1.setMaximum(100)
-        self.slider_DAC1.setValue(0)
-        self.slider_DAC1.setSingleStep(0.1)
-
-        self.entry_DAC1 = QDoubleSpinBox()
-        self.entry_DAC1.setMinimum(0) 
-        self.entry_DAC1.setMaximum(100) 
-        self.entry_DAC1.setSingleStep(0.1)
-        self.entry_DAC1.setValue(0)
-        self.entry_DAC1.setKeyboardTracking(False)
-
-        # connections
-        self.entry_DAC0.valueChanged.connect(self.set_DAC0)
-        self.entry_DAC0.valueChanged.connect(self.slider_DAC0.setValue)
-        self.slider_DAC0.valueChanged.connect(self.entry_DAC0.setValue)
-        self.entry_DAC1.valueChanged.connect(self.set_DAC1)
-        self.entry_DAC1.valueChanged.connect(self.slider_DAC1.setValue)
-        self.slider_DAC1.valueChanged.connect(self.entry_DAC1.setValue)
-
-        # layout
-        grid_line1 = QGridLayout()
-        grid_line1.addWidget(QLabel('DAC0'), 0,0)
-        grid_line1.addWidget(self.slider_DAC0, 0,1)
-        grid_line1.addWidget(self.entry_DAC0, 0,2)
-        grid_line1.addWidget(QLabel('DAC1'), 1,0)
-        grid_line1.addWidget(self.slider_DAC1, 1,1)
-        grid_line1.addWidget(self.entry_DAC1, 1,2)
-
-        self.grid = QGridLayout()
-        self.grid.addLayout(grid_line1,1,0)
-        self.setLayout(self.grid)
-
-    def set_DAC0(self,value):
-        self.microcontroller.analog_write_onboard_DAC(0,int(value*65535/100))
-
-    def set_DAC1(self,value):
-        self.microcontroller.analog_write_onboard_DAC(1,int(value*65535/100))
-
-class AutoFocusWidget(QFrame):
-    def __init__(self, autofocusController, main=None, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.autofocusController = autofocusController
-        self.add_components()
-        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
-
-    def add_components(self):
-        self.entry_delta = QDoubleSpinBox()
-        self.entry_delta.setMinimum(0) 
-        self.entry_delta.setMaximum(20) 
-        self.entry_delta.setSingleStep(0.2)
-        self.entry_delta.setDecimals(3)
-        self.entry_delta.setValue(1.524)
-        self.entry_delta.setKeyboardTracking(False)
-        self.autofocusController.set_deltaZ(1.524)
-
-        self.entry_N = QSpinBox()
-        self.entry_N.setMinimum(3) 
-        self.entry_N.setMaximum(20) 
-        self.entry_N.setSingleStep(1)
-        self.entry_N.setValue(10)
-        self.entry_N.setKeyboardTracking(False)
-        self.autofocusController.set_N(10)
-
-        self.btn_autofocus = QPushButton('Autofocus')
-        self.btn_autofocus.setDefault(False)
-        self.btn_autofocus.setCheckable(True)
-        self.btn_autofocus.setChecked(False)
-
-        # layout
-        grid_line0 = QGridLayout()
-        dz_tooltip="use autofocus by taking z-stack of images (NZ images, with dz um distance between images), then \ncalculating a focus metric and choosing the image plane with the best metric.\n\nthe images are taken in the channel that is currently selected for live view (led+micro will be turned on if they are off)\n\nthis will take a few seconds"
-        self.btn_autofocus.setToolTip(dz_tooltip)
-        qtlabel_dz=QLabel('delta Z (um)')
-        qtlabel_dz.setToolTip(dz_tooltip)
-        grid_line0.addWidget(qtlabel_dz, 0,0)
-        grid_line0.addWidget(self.entry_delta, 0,1)
-        qtlabel_Nz=QLabel('N Z planes')
-        qtlabel_Nz.setToolTip(dz_tooltip)
-        grid_line0.addWidget(qtlabel_Nz, 0,2)
-        grid_line0.addWidget(self.entry_N, 0,3)
-        grid_line0.addWidget(self.btn_autofocus, 0,4)
-
-        self.grid = QGridLayout()
-        self.grid.addLayout(grid_line0,0,0)
-        self.setLayout(self.grid)
-        
-        # connections
-        self.btn_autofocus.clicked.connect(self.autofocusController.autofocus)
-        self.entry_delta.valueChanged.connect(self.set_deltaZ)
-        self.entry_N.valueChanged.connect(self.autofocusController.set_N)
-        self.autofocusController.autofocusFinished.connect(self.autofocus_is_finished)
-
-    def set_deltaZ(self,value):
-        mm_per_ustep = SCREW_PITCH_Z_MM/(self.autofocusController.navigationController.z_microstepping*FULLSTEPS_PER_REV_Z)
-        deltaZ = round(value/1000/mm_per_ustep)*mm_per_ustep*1000
-        self.entry_delta.setValue(deltaZ)
-        self.autofocusController.set_deltaZ(deltaZ)
-
-    def autofocus_is_finished(self):
-        self.btn_autofocus.setChecked(False)
-
-autofocus_channel_names=[
-    "Fluorescence 405 nm Ex",
-    "Fluorescence 488 nm Ex",
-    "Fluorescence 561 nm Ex",
-    "Fluorescence 638 nm Ex",
-    "Fluorescence 730 nm Ex",
-]
-
-class MultiPointWidget(QFrame):
-    def __init__(self, multipointController, configurationManager = None, main=None, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.multipointController = multipointController
-        self.configurationManager = configurationManager
-        self.base_path_is_set = False
-        self.add_components()
-        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
-
-    def add_components(self):
-
-        self.btn_setSavingDir = QPushButton('Browse')
-        self.btn_setSavingDir.setDefault(False)
-        self.btn_setSavingDir.setIcon(QIcon('icon/folder.png'))
-        
-        self.lineEdit_savingDir = QLineEdit()
-        self.lineEdit_savingDir.setReadOnly(True)
-        self.lineEdit_savingDir.setText('Choose a base saving directory')
-
-        self.lineEdit_savingDir.setText(DEFAULT_SAVING_PATH)
-        self.multipointController.set_base_path(DEFAULT_SAVING_PATH)
-        self.base_path_is_set = True
-
-        self.lineEdit_experimentID = QLineEdit()
-
-        self.entry_deltaX = QDoubleSpinBox()
-        self.entry_deltaX.setMinimum(0) 
-        self.entry_deltaX.setMaximum(5) 
-        self.entry_deltaX.setSingleStep(0.1)
-        self.entry_deltaX.setValue(Acquisition.DX)
-        self.entry_deltaX.setDecimals(3)
-        self.entry_deltaX.setKeyboardTracking(False)
-
-        self.entry_NX = QSpinBox()
-        self.entry_NX.setMinimum(1) 
-        self.entry_NX.setMaximum(50) 
-        self.entry_NX.setSingleStep(1)
-        self.entry_NX.setValue(1)
-        self.entry_NX.setKeyboardTracking(False)
-
-        self.entry_deltaY = QDoubleSpinBox()
-        self.entry_deltaY.setMinimum(0) 
-        self.entry_deltaY.setMaximum(5) 
-        self.entry_deltaY.setSingleStep(0.1)
-        self.entry_deltaY.setValue(Acquisition.DX)
-        self.entry_deltaY.setDecimals(3)
-        self.entry_deltaY.setKeyboardTracking(False)
-        
-        self.entry_NY = QSpinBox()
-        self.entry_NY.setMinimum(1) 
-        self.entry_NY.setMaximum(50) 
-        self.entry_NY.setSingleStep(1)
-        self.entry_NY.setValue(1)
-        self.entry_NY.setKeyboardTracking(False)
-
-        self.entry_deltaZ = QDoubleSpinBox()
-        self.entry_deltaZ.setMinimum(0) 
-        self.entry_deltaZ.setMaximum(1000) 
-        self.entry_deltaZ.setSingleStep(0.2)
-        self.entry_deltaZ.setValue(Acquisition.DZ)
-        self.entry_deltaZ.setDecimals(3)
-        self.entry_deltaZ.setKeyboardTracking(False)
-        
-        self.entry_NZ = QSpinBox()
-        self.entry_NZ.setMinimum(1) 
-        self.entry_NZ.setMaximum(100) 
-        self.entry_NZ.setSingleStep(1)
-        self.entry_NZ.setValue(1)
-        self.entry_NZ.setKeyboardTracking(False)
-        
-        self.entry_dt = QDoubleSpinBox()
-        self.entry_dt.setMinimum(0) 
-        self.entry_dt.setMaximum(12*3600) 
-        self.entry_dt.setSingleStep(1)
-        self.entry_dt.setValue(0)
-        self.entry_dt.setKeyboardTracking(False)
-
-        self.entry_Nt = QSpinBox()
-        self.entry_Nt.setMinimum(1) 
-        self.entry_Nt.setMaximum(50000)   # @@@ to be changed
-        self.entry_Nt.setSingleStep(1)
-        self.entry_Nt.setValue(1)
-        self.entry_Nt.setKeyboardTracking(False)
-
-        self.list_configurations = QListWidget()
-        for microscope_configuration in self.configurationManager.configurations:
-            self.list_configurations.addItems([microscope_configuration.name])
-        self.list_configurations.setSelectionMode(QAbstractItemView.MultiSelection) # ref: https://doc.qt.io/qt-5/qabstractitemview.html#SelectionMode-enum
-
-        self.checkbox_withAutofocus = QCheckBox('With AF')
-        self.checkbox_withAutofocus.setToolTip("enable autofocus for multipoint acquisition\nfor each well the autofocus will be calculated in the channel selected below")
-        self.checkbox_withAutofocus.setChecked(MULTIPOINT_AUTOFOCUS_ENABLE_BY_DEFAULT)
-        self.multipointController.set_af_flag(MULTIPOINT_AUTOFOCUS_ENABLE_BY_DEFAULT)
-        self.btn_startAcquisition = QPushButton('Start Acquisition')
-        self.btn_startAcquisition.setCheckable(True)
-        self.btn_startAcquisition.setChecked(False)
-
-        # layout
-        grid_line0 = QGridLayout()
-        grid_line0.addWidget(QLabel('Saving Path'))
-        grid_line0.addWidget(self.lineEdit_savingDir, 0,1)
-        grid_line0.addWidget(self.btn_setSavingDir, 0,2)
-
-        grid_line1 = QGridLayout()
-        grid_line1.addWidget(QLabel('Experiment ID'), 0,0)
-        grid_line1.addWidget(self.lineEdit_experimentID,0,1)
-
-        grid_line2 = QGridLayout()
-        dx_tooltip="acquire grid of images (Nx images with dx mm in between acquisitions; dx does not matter if Nx is 1)\ncan be combined with dy/Ny and dz/Nz and dt/Nt for a total of Nx * Ny * Nz * Nt images"
-        qtlabel_dx=QLabel('dx (mm)')
-        qtlabel_dx.setToolTip(dx_tooltip)
-        grid_line2.addWidget(qtlabel_dx, 0,0)
-        grid_line2.addWidget(self.entry_deltaX, 0,1)
-        qtlabel_Nx=QLabel('Nx')
-        qtlabel_Nx.setToolTip(dx_tooltip)
-        grid_line2.addWidget(qtlabel_Nx, 0,2)
-        grid_line2.addWidget(self.entry_NX, 0,3)
- 
-        dy_tooltip="acquire grid of images (Ny images with dy mm in between acquisitions; dy does not matter if Ny is 1)\ncan be combined with dx/Nx and dz/Nz and dt/Nt for a total of Nx*Ny*Nz*Nt images"
-        qtlabel_dy=QLabel('dy (mm)')
-        qtlabel_dy.setToolTip(dy_tooltip)
-        grid_line2.addWidget(qtlabel_dy, 1,0)
-        grid_line2.addWidget(self.entry_deltaY, 1,1)
-        qtlabel_Ny=QLabel('Ny')
-        qtlabel_Ny.setToolTip(dy_tooltip)
-        grid_line2.addWidget(qtlabel_Ny, 1,2)
-        grid_line2.addWidget(self.entry_NY, 1,3)
- 
-        dz_tooltip="acquire z-stack of images (Nz images with dz µm in between acquisitions; dz does not matter if Nz is 1)\ncan be combined with dx/Nx and dy/Ny and dt/Nt for a total of Nx*Ny*Nz*Nt images"
-        qtlabel_dz=QLabel('dz (um)')
-        qtlabel_dz.setToolTip(dx_tooltip)
-        grid_line2.addWidget(qtlabel_dz, 2,0)
-        grid_line2.addWidget(self.entry_deltaZ, 2,1)
-        qtlabel_Nz=QLabel('Nz')
-        qtlabel_Nz.setToolTip(dx_tooltip)
-        grid_line2.addWidget(qtlabel_Nz, 2,2)
-        grid_line2.addWidget(self.entry_NZ, 2,3)
- 
-        dt_tooltip="acquire time-series of 'Nt' images, with 'dt' seconds in between acquisitions (dt does not matter if Nt is 1)\ncan be combined with dx/Nx and dy/Ny and dz/Nz for a total of Nx*Ny*Nz*Nt images"
-        qtlabel_dt=QLabel('dt (s)')
-        qtlabel_dt.setToolTip(dt_tooltip)
-        grid_line2.addWidget(qtlabel_dt, 3,0)
-        grid_line2.addWidget(self.entry_dt, 3,1)
-        qtlabel_Nt=QLabel('Nt')
-        qtlabel_Nt.setToolTip(dt_tooltip)
-        grid_line2.addWidget(qtlabel_Nt, 3,2)
-        grid_line2.addWidget(self.entry_Nt, 3,3)
-
-        af_channel_dropdown=QComboBox()
-        af_channel_dropdown.setToolTip("set channel that will be used for autofocus measurements")
-        af_channel_dropdown.addItems(autofocus_channel_names)
-        af_channel_dropdown.setCurrentIndex(autofocus_channel_names.index(self.multipointController.autofocus_channel_name))
-        af_channel_dropdown.currentIndexChanged.connect(self.set_autofocusChannel)
-
-        grid_multipoint_acquisition_config=QGridLayout()
-        grid_multipoint_acquisition_config.addWidget(self.checkbox_withAutofocus,0,0)
-        grid_multipoint_acquisition_config.addWidget(af_channel_dropdown,1,0)
-        grid_multipoint_acquisition_config.addWidget(self.btn_startAcquisition,2,0)
-
-        grid_line3 = QHBoxLayout()
-        grid_line3.addWidget(self.list_configurations)
-        grid_line3.addLayout(grid_multipoint_acquisition_config)
-
-        self.grid = QGridLayout()
-        self.grid.addLayout(grid_line0,0,0)
-        self.grid.addLayout(grid_line1,1,0)
-        self.grid.addLayout(grid_line2,2,0)
-        self.grid.addLayout(grid_line3,3,0)
-        self.setLayout(self.grid)
-
-        # add and display a timer - to be implemented
-        # self.timer = QTimer()
-
-        # connections
-        self.entry_deltaX.valueChanged.connect(self.set_deltaX)
-        self.entry_deltaY.valueChanged.connect(self.set_deltaY)
-        self.entry_deltaZ.valueChanged.connect(self.set_deltaZ)
-        self.entry_dt.valueChanged.connect(self.multipointController.set_deltat)
-        self.entry_NX.valueChanged.connect(self.multipointController.set_NX)
-        self.entry_NY.valueChanged.connect(self.multipointController.set_NY)
-        self.entry_NZ.valueChanged.connect(self.multipointController.set_NZ)
-        self.entry_Nt.valueChanged.connect(self.multipointController.set_Nt)
-        self.checkbox_withAutofocus.stateChanged.connect(self.multipointController.set_af_flag)
-        self.btn_setSavingDir.clicked.connect(self.set_saving_dir)
-        self.btn_startAcquisition.clicked.connect(self.toggle_acquisition)
-        self.multipointController.acquisitionFinished.connect(self.acquisition_is_finished)
-
-    def set_autofocusChannel(self,new_channel_index):
-        self.multipointController.autofocus_channel_name=autofocus_channel_names[new_channel_index]
-
-    def set_deltaX(self,value):
-        mm_per_ustep = SCREW_PITCH_X_MM/(self.multipointController.navigationController.x_microstepping*FULLSTEPS_PER_REV_X) # to implement a get_x_microstepping() in multipointController
-        deltaX = round(value/mm_per_ustep)*mm_per_ustep
-        self.entry_deltaX.setValue(deltaX)
-        self.multipointController.set_deltaX(deltaX)
-
-    def set_deltaY(self,value):
-        mm_per_ustep = SCREW_PITCH_Y_MM/(self.multipointController.navigationController.y_microstepping*FULLSTEPS_PER_REV_Y)
-        deltaY = round(value/mm_per_ustep)*mm_per_ustep
-        self.entry_deltaY.setValue(deltaY)
-        self.multipointController.set_deltaY(deltaY)
-
-    def set_deltaZ(self,value):
-        mm_per_ustep = SCREW_PITCH_Z_MM/(self.multipointController.navigationController.z_microstepping*FULLSTEPS_PER_REV_Z)
-        deltaZ = round(value/1000/mm_per_ustep)*mm_per_ustep*1000
-        self.entry_deltaZ.setValue(deltaZ)
-        self.multipointController.set_deltaZ(deltaZ)
-
-    def set_saving_dir(self):
-        dialog = QFileDialog()
-        save_dir_base = dialog.getExistingDirectory(None, "Select Folder")
-        self.multipointController.set_base_path(save_dir_base)
-        self.lineEdit_savingDir.setText(save_dir_base)
-        self.base_path_is_set = True
-
-    def toggle_acquisition(self,pressed):
-        if self.base_path_is_set == False:
-            self.btn_startAcquisition.setChecked(False)
-            msg = QMessageBox()
-            msg.setText("Please choose base saving directory first")
-            msg.exec_()
-            return
-        if pressed:
-            # @@@ to do: add a widgetManger to enable and disable widget 
-            # @@@ to do: emit signal to widgetManager to disable other widgets
-            self.setEnabled_all(False)
-            self.multipointController.start_new_experiment(self.lineEdit_experimentID.text())
-            self.multipointController.set_selected_configurations((item.text() for item in self.list_configurations.selectedItems()))
-            self.multipointController.run_experiment()
-        else:
-            self.multipointController.request_abort_aquisition()
-            self.setEnabled_all(True)
-
-    def acquisition_is_finished(self):
-        self.btn_startAcquisition.setChecked(False)
-        self.setEnabled_all(True)
-
-    def setEnabled_all(self,enabled,exclude_btn_startAcquisition=True):
-        self.btn_setSavingDir.setEnabled(enabled)
-        self.lineEdit_savingDir.setEnabled(enabled)
-        self.lineEdit_experimentID.setEnabled(enabled)
-        self.entry_deltaX.setEnabled(enabled)
-        self.entry_NX.setEnabled(enabled)
-        self.entry_deltaY.setEnabled(enabled)
-        self.entry_NY.setEnabled(enabled)
-        self.entry_deltaZ.setEnabled(enabled)
-        self.entry_NZ.setEnabled(enabled)
-        self.entry_dt.setEnabled(enabled)
-        self.entry_Nt.setEnabled(enabled)
-        self.list_configurations.setEnabled(enabled)
-        self.checkbox_withAutofocus.setEnabled(enabled)
-        if exclude_btn_startAcquisition is not True:
-            self.btn_startAcquisition.setEnabled(enabled)
-
-    def disable_the_start_aquisition_button(self):
-        self.btn_startAcquisition.setEnabled(False)
-
-    def enable_the_start_aquisition_button(self):
-        self.btn_startAcquisition.setEnabled(True)
-
-class WellSelectionWidget(QTableWidget):
- 
-    signal_wellSelected = Signal(int,int,float)
-    signal_wellSelectedPos = Signal(float,float)
- 
-    def __init__(self, format: int):
-        self.was_initialized=False
-        self.set_wellplate_type(format)
-        self.was_initialized=True
- 
-    def set_wellplate_type(self,wellplate_type:Union[str,int]):
-        if type(wellplate_type)==str:
-            wellplate_type_int:int=int(wellplate_type.split(" ")[0])  # type: ignore
-        else:
-            wellplate_type_int:int=wellplate_type # type: ignore
- 
-        wellplate_type_format=WELLPLATE_FORMATS[wellplate_type_int]
-        self.rows = wellplate_type_format.rows
-        self.columns = wellplate_type_format.columns
-        self.spacing_mm = wellplate_type_format.well_spacing_mm
- 
-        if self.was_initialized:
-            old_layout=WELLPLATE_FORMATS[self.format]
-            self.set_selectable_widgets(layout=old_layout,is_selectable=True,exhaustive=True)
- 
-            self.format:int=wellplate_type_int
- 
-            self.setRowCount(self.rows)
-            self.setColumnCount(self.columns)
- 
-            self.setData()
-        else:
-            self.format=wellplate_type_int
- 
-            QTableWidget.__init__(self, self.rows, self.columns)
- 
-            self.setData()
- 
-        self.resizeColumnsToContents()
-        self.resizeRowsToContents()
-        if not self.was_initialized:
-            self.setEditTriggers(QTableWidget.NoEditTriggers)
-            self.cellDoubleClicked.connect(self.onDoubleClick)
- 
-        # size
-        self.verticalHeader().setSectionResizeMode(QHeaderView.Fixed)
-        self.verticalHeader().setDefaultSectionSize(5*self.spacing_mm)
-        self.horizontalHeader().setSectionResizeMode(QHeaderView.Fixed)
-        self.horizontalHeader().setMinimumSectionSize(5*self.spacing_mm)
- 
-        self.setSizePolicy(QSizePolicy.Minimum,QSizePolicy.Minimum)
-        self.setVerticalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.resizeColumnsToContents()
-        self.setFixedSize(
-            self.horizontalHeader().length() + self.verticalHeader().width(),
-            self.verticalHeader().length() + self.horizontalHeader().height()
-        )
- 
-    def set_selectable_widgets(self,layout:WellplateFormatPhysical,is_selectable:bool,exhaustive:bool=False):
-        # item.flags is a bitvector, so changing the IsSelectable flag is bit manipulating magic
- 
-        if not is_selectable:
-            assert not exhaustive, "cannot exhaustively disable only outer ring"
- 
-        for i in range(layout.rows):
-            for j in (range(layout.columns) if exhaustive else [0,layout.columns-1]):
-                item = QTableWidgetItem()
-                item.setFlags((item.flags() | Qt.ItemIsSelectable) if is_selectable else (item.flags() & ~Qt.ItemIsSelectable))
-                self.setItem(i,j,item)
- 
-        if not exhaustive:
-            for j in range(layout.columns):
-                for i in [0,layout.rows-1]:
-                    item = QTableWidgetItem()
-                    item.setFlags((item.flags() | Qt.ItemIsSelectable) if is_selectable else (item.flags() & ~Qt.ItemIsSelectable))
-                    self.setItem(i,j,item)
- 
-    def setData(self):
-        '''
-        # cells
-        for i in range(16):
-            for j in range(24):
-                newitem = QTableWidgetItem( chr(ord('A')+i) + str(j) )
-                self.setItem(i, j, newitem)
-        '''
-        # row header
-        row_headers = []
-        for i in range(self.rows):
-            row_headers.append(chr(ord('A')+i))
-        self.setVerticalHeaderLabels(row_headers)
- 
-        # make the outer cells not selectable if using 96 and 384 well plates
-        wellplate_format=WELLPLATE_FORMATS[self.format]
- 
-        if wellplate_format.number_of_skip==1:
-            self.set_selectable_widgets(layout=wellplate_format,is_selectable=False)
-        elif wellplate_format.number_of_skip>1:
-            assert False, "more than one layer of disabled outer wells is currently unimplemented"
- 
-    def onDoubleClick(self,row:int,col:int):
-        wellplate_format=WELLPLATE_FORMATS[self.format]
- 
-        if (row >= 0 + wellplate_format.number_of_skip and row <= self.rows-1-wellplate_format.number_of_skip ) and ( col >= 0 + wellplate_format.number_of_skip and col <= self.columns-1-wellplate_format.number_of_skip ):
-            wellplateformat_384=WELLPLATE_FORMATS[384]
-            x_mm = X_MM_384_WELLPLATE_UPPERLEFT + wellplateformat_384.well_size_mm/2 - (wellplateformat_384.A1_x_mm+wellplateformat_384.well_spacing_mm*wellplateformat_384.number_of_skip) + col*wellplate_format.well_spacing_mm + wellplate_format.A1_x_mm + WELLPLATE_OFFSET_X_mm
-            y_mm = Y_MM_384_WELLPLATE_UPPERLEFT + wellplateformat_384.well_size_mm/2 - (wellplateformat_384.A1_y_mm+wellplateformat_384.well_spacing_mm*wellplateformat_384.number_of_skip) + row*wellplate_format.well_spacing_mm + wellplate_format.A1_y_mm + WELLPLATE_OFFSET_Y_mm
-            self.signal_wellSelectedPos.emit(x_mm,y_mm)
- 
-    def get_selected_cells(self) -> List[Tuple[int,int]]:
-        list_of_selected_cells = []
-        for index in self.selectedIndexes():
-             list_of_selected_cells.append((index.row(),index.column()))
-        return list_of_selected_cells
