+++ software/control/widgets/__init__.py
@@ -1,0 +1,12 @@
+# set QT_API environment variable
+import os 
+os.environ["QT_API"] = "pyqt5"
+
+from .autofocus import AutoFocusWidget
+from .camera_settings import CameraSettingsWidget
+from .dac_control import DACControWidget
+from .live_control import LiveControlWidget
+from .multi_point import MultiPointWidget
+from .navigation import NavigationWidget
+from .recording import RecordingWidget
+from .well_selection import WellSelectionWidget
