+++ software/control/widgets/dac_control.py
@@ -1,0 +1,76 @@
+# qt libraries
+from qtpy.QtCore import *
+from qtpy.QtWidgets import *
+from qtpy.QtGui import *
+
+import pyqtgraph as pg
+
+from datetime import datetime
+
+from control._def import *
+import control.core as core
+
+from typing import Optional, Union, List, Tuple
+
+class DACControWidget(QFrame):
+    def __init__(self, microcontroller ,*args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.microcontroller = microcontroller
+        self.add_components()
+        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
+
+    def add_components(self):
+        self.slider_DAC0 = QSlider(Qt.Horizontal)
+        self.slider_DAC0.setTickPosition(QSlider.TicksBelow)
+        self.slider_DAC0.setMinimum(0)
+        self.slider_DAC0.setMaximum(100)
+        self.slider_DAC0.setSingleStep(0.1)
+        self.slider_DAC0.setValue(0)
+
+        self.entry_DAC0 = QDoubleSpinBox()
+        self.entry_DAC0.setMinimum(0) 
+        self.entry_DAC0.setMaximum(100) 
+        self.entry_DAC0.setSingleStep(0.1)
+        self.entry_DAC0.setValue(0)
+        self.entry_DAC0.setKeyboardTracking(False)
+
+        self.slider_DAC1 = QSlider(Qt.Horizontal)
+        self.slider_DAC1.setTickPosition(QSlider.TicksBelow)
+        self.slider_DAC1.setMinimum(0)
+        self.slider_DAC1.setMaximum(100)
+        self.slider_DAC1.setValue(0)
+        self.slider_DAC1.setSingleStep(0.1)
+
+        self.entry_DAC1 = QDoubleSpinBox()
+        self.entry_DAC1.setMinimum(0) 
+        self.entry_DAC1.setMaximum(100) 
+        self.entry_DAC1.setSingleStep(0.1)
+        self.entry_DAC1.setValue(0)
+        self.entry_DAC1.setKeyboardTracking(False)
+
+        # connections
+        self.entry_DAC0.valueChanged.connect(self.set_DAC0)
+        self.entry_DAC0.valueChanged.connect(self.slider_DAC0.setValue)
+        self.slider_DAC0.valueChanged.connect(self.entry_DAC0.setValue)
+        self.entry_DAC1.valueChanged.connect(self.set_DAC1)
+        self.entry_DAC1.valueChanged.connect(self.slider_DAC1.setValue)
+        self.slider_DAC1.valueChanged.connect(self.entry_DAC1.setValue)
+
+        # layout
+        grid_line1 = QGridLayout()
+        grid_line1.addWidget(QLabel('DAC0'), 0,0)
+        grid_line1.addWidget(self.slider_DAC0, 0,1)
+        grid_line1.addWidget(self.entry_DAC0, 0,2)
+        grid_line1.addWidget(QLabel('DAC1'), 1,0)
+        grid_line1.addWidget(self.slider_DAC1, 1,1)
+        grid_line1.addWidget(self.entry_DAC1, 1,2)
+
+        self.grid = QGridLayout()
+        self.grid.addLayout(grid_line1,1,0)
+        self.setLayout(self.grid)
+
+    def set_DAC0(self,value):
+        self.microcontroller.analog_write_onboard_DAC(0,int(value*65535/100))
+
+    def set_DAC1(self,value):
+        self.microcontroller.analog_write_onboard_DAC(1,int(value*65535/100))
