+++ software/control/widgets/live_control.py
@@ -1,0 +1,234 @@
+# qt libraries
+from qtpy.QtCore import *
+from qtpy.QtWidgets import *
+from qtpy.QtGui import *
+
+import pyqtgraph as pg
+
+from datetime import datetime
+
+from control._def import *
+import control.core as core
+
+from typing import Optional, Union, List, Tuple
+# 'Live' button text
+LIVE_BUTTON_IDLE_TEXT="Start Live"
+LIVE_BUTTON_RUNNING_TEXT="Stop Live"
+
+class LiveControlWidget(QFrame):
+    signal_newExposureTime = Signal(float)
+    signal_newAnalogGain = Signal(float)
+    signal_autoLevelSetting = Signal(bool)
+
+    def __init__(self, 
+        streamHandler, 
+        liveController:core.LiveController,
+        configurationManager:core.ConfigurationManager, 
+        show_trigger_options:bool=True, 
+        show_display_options:bool=True, 
+        show_autolevel:bool = False, 
+        autolevel:bool=False
+    ):
+        super().__init__()
+        self.liveController = liveController
+        self.streamHandler = streamHandler
+        self.configurationManager = configurationManager
+        self.fps_trigger:float = 10
+        self.fps_display:float = 10
+        self.liveController.set_trigger_fps(self.fps_trigger)
+        self.streamHandler.set_display_fps(self.fps_display)
+        
+        self.triggerMode = TriggerMode.SOFTWARE
+        # note that this references the object in self.configurationManager.configurations
+        self.currentConfiguration = self.configurationManager.configurations[0]
+
+        self.add_components(show_trigger_options,show_display_options,show_autolevel,autolevel)
+        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
+        self.update_microscope_mode_by_name(self.currentConfiguration.name)
+
+        self.is_switching_mode = False # flag used to prevent from settings being set by twice - from both mode change slot and value change slot; another way is to use blockSignals(True)
+
+    def add_components(self,
+        show_trigger_options:bool,
+        show_display_options:bool,
+        show_autolevel:bool,
+        autolevel:bool
+    ):
+        # line 0: trigger mode
+        self.triggerMode = None
+        self.dropdown_triggerManu = QComboBox()
+        self.dropdown_triggerManu.addItems([TriggerMode.SOFTWARE,TriggerMode.HARDWARE,TriggerMode.CONTINUOUS])
+
+        # line 1: fps
+        self.entry_triggerFPS = QDoubleSpinBox()
+        self.entry_triggerFPS.setMinimum(0.02) 
+        self.entry_triggerFPS.setMaximum(1000) 
+        self.entry_triggerFPS.setSingleStep(1)
+        self.entry_triggerFPS.setValue(self.fps_trigger)
+
+        # line 2: choose microscope mode / toggle live mode 
+        self.dropdown_modeSelection = QComboBox()
+        for microscope_configuration in self.configurationManager.configurations:
+            self.dropdown_modeSelection.addItems([microscope_configuration.name])
+        self.dropdown_modeSelection.setCurrentText(self.currentConfiguration.name)
+
+        self.btn_live = QPushButton(LIVE_BUTTON_IDLE_TEXT)
+        self.btn_live.setCheckable(True)
+        self.btn_live.setChecked(False)
+        self.btn_live.setDefault(False)
+
+        # line 3: exposure time and analog gain associated with the current mode
+        self.entry_exposureTime = QDoubleSpinBox()
+        self.entry_exposureTime.setMinimum(self.liveController.camera.EXPOSURE_TIME_MS_MIN) 
+        self.entry_exposureTime.setMaximum(self.liveController.camera.EXPOSURE_TIME_MS_MAX) 
+        self.entry_exposureTime.setSingleStep(1)
+        self.entry_exposureTime.setValue(0)
+
+        self.entry_analogGain = QDoubleSpinBox()
+        self.entry_analogGain = QDoubleSpinBox()
+        self.entry_analogGain.setMinimum(0) 
+        self.entry_analogGain.setMaximum(24) 
+        self.entry_analogGain.setSingleStep(0.1)
+        self.entry_analogGain.setValue(0)
+
+        self.slider_illuminationIntensity = QSlider(Qt.Horizontal)
+        self.slider_illuminationIntensity.setTickPosition(QSlider.TicksBelow)
+        self.slider_illuminationIntensity.setMinimum(0)
+        self.slider_illuminationIntensity.setMaximum(100)
+        self.slider_illuminationIntensity.setValue(100)
+        self.slider_illuminationIntensity.setSingleStep(0.1)
+
+        self.entry_illuminationIntensity = QDoubleSpinBox()
+        self.entry_illuminationIntensity.setMinimum(0.1) 
+        self.entry_illuminationIntensity.setMaximum(100) 
+        self.entry_illuminationIntensity.setSingleStep(0.1)
+        self.entry_illuminationIntensity.setValue(100)
+
+        # line 4: display fps and resolution scaling
+        self.entry_displayFPS = QDoubleSpinBox()
+        self.entry_displayFPS.setMinimum(1) 
+        self.entry_displayFPS.setMaximum(240) 
+        self.entry_displayFPS.setSingleStep(1)
+        self.entry_displayFPS.setValue(self.fps_display)
+
+        self.slider_resolutionScaling = QSlider(Qt.Horizontal)
+        self.slider_resolutionScaling.setTickPosition(QSlider.TicksBelow)
+        self.slider_resolutionScaling.setMinimum(10)
+        self.slider_resolutionScaling.setMaximum(100)
+        self.slider_resolutionScaling.setValue(DEFAULT_DISPLAY_CROP)
+        self.slider_resolutionScaling.setSingleStep(10)
+
+        # autolevel
+        self.btn_autolevel = QPushButton('Autolevel')
+        self.btn_autolevel.setCheckable(True)
+        self.btn_autolevel.setChecked(autolevel)
+        
+        # connections
+        self.entry_triggerFPS.valueChanged.connect(self.liveController.set_trigger_fps)
+        self.entry_displayFPS.valueChanged.connect(self.streamHandler.set_display_fps)
+        self.slider_resolutionScaling.valueChanged.connect(self.streamHandler.set_display_resolution_scaling)
+        self.slider_resolutionScaling.valueChanged.connect(self.liveController.set_display_resolution_scaling)
+        self.dropdown_modeSelection.currentTextChanged.connect(self.update_microscope_mode_by_name)
+        self.dropdown_triggerManu.currentIndexChanged.connect(self.update_trigger_mode)
+        self.btn_live.clicked.connect(self.toggle_live)
+        self.entry_exposureTime.valueChanged.connect(self.update_config_exposure_time)
+        self.entry_analogGain.valueChanged.connect(self.update_config_analog_gain)
+        self.entry_illuminationIntensity.valueChanged.connect(self.update_config_illumination_intensity)
+        self.entry_illuminationIntensity.valueChanged.connect(self.slider_illuminationIntensity.setValue)
+        self.slider_illuminationIntensity.valueChanged.connect(self.entry_illuminationIntensity.setValue)
+        self.btn_autolevel.clicked.connect(self.signal_autoLevelSetting.emit)
+
+        # layout
+        grid_line0 = QGridLayout()
+        grid_line0.addWidget(QLabel('Trigger Mode'), 0,0)
+        grid_line0.addWidget(self.dropdown_triggerManu, 0,1)
+        grid_line0.addWidget(QLabel('Trigger FPS'), 0,2)
+        grid_line0.addWidget(self.entry_triggerFPS, 0,3)
+
+        grid_line1 = QGridLayout()
+        grid_line1.addWidget(QLabel('Imaging Mode'), 0,0)
+        grid_line1.addWidget(self.dropdown_modeSelection, 0,1)
+        grid_line1.addWidget(self.btn_live, 0,2)
+
+        grid_line2 = QGridLayout()
+        grid_line2.addWidget(QLabel('Exposure Time (ms)'), 0,0)
+        grid_line2.addWidget(self.entry_exposureTime, 0,1)
+        grid_line2.addWidget(QLabel('Analog Gain'), 0,2)
+        grid_line2.addWidget(self.entry_analogGain, 0,3)
+
+        grid_line4 = QGridLayout()
+        grid_line4.addWidget(QLabel('Illumination'), 0,0)
+        grid_line4.addWidget(self.slider_illuminationIntensity, 0,1)
+        grid_line4.addWidget(self.entry_illuminationIntensity, 0,2)
+
+        grid_line3 = QGridLayout()
+        grid_line3.addWidget(QLabel('Display FPS'), 0,0)
+        grid_line3.addWidget(self.entry_displayFPS, 0,1)
+        grid_line3.addWidget(QLabel('Display Resolution'), 0,2)
+        grid_line3.addWidget(self.slider_resolutionScaling,0,3)
+        if show_autolevel:
+            grid_line3.addWidget(self.btn_autolevel,0,4)
+
+        self.grid = QVBoxLayout()
+        if show_trigger_options:
+            self.grid.addLayout(grid_line0)
+        self.grid.addLayout(grid_line1)
+        self.grid.addLayout(grid_line2)
+        self.grid.addLayout(grid_line4)
+        if show_display_options:
+            self.grid.addLayout(grid_line3)
+        self.grid.addStretch()
+        self.setLayout(self.grid)
+
+    def toggle_live(self,pressed:bool):
+        if pressed:
+            self.btn_live.setText(LIVE_BUTTON_RUNNING_TEXT)
+            self.liveController.start_live()
+        else:
+            self.btn_live.setText(LIVE_BUTTON_IDLE_TEXT)
+            self.liveController.stop_live()
+
+    def update_camera_settings(self):
+        self.signal_newAnalogGain.emit(self.entry_analogGain.value())
+        self.signal_newExposureTime.emit(self.entry_exposureTime.value())
+
+    def update_microscope_mode_by_name(self,current_microscope_mode_name:str):
+        self.is_switching_mode = True
+        # identify the mode selected (note that this references the object in self.configurationManager.configurations)
+        self.currentConfiguration = next((config for config in self.configurationManager.configurations if config.name == current_microscope_mode_name), None)
+        # update the microscope to the current configuration
+        self.liveController.set_microscope_mode(self.currentConfiguration)
+        # update the exposure time and analog gain settings according to the selected configuration
+        self.entry_exposureTime.setValue(self.currentConfiguration.exposure_time)
+        self.entry_analogGain.setValue(self.currentConfiguration.analog_gain)
+        self.entry_illuminationIntensity.setValue(self.currentConfiguration.illumination_intensity)
+        self.is_switching_mode = False
+
+    def update_trigger_mode(self):
+        self.liveController.set_trigger_mode(self.dropdown_triggerManu.currentText())
+
+    def update_config_exposure_time(self,new_value:float):
+        if self.is_switching_mode == False:
+            self.currentConfiguration.exposure_time = new_value
+            self.configurationManager.update_configuration(self.currentConfiguration.id,'ExposureTime',new_value)
+            self.signal_newExposureTime.emit(new_value)
+
+    def update_config_analog_gain(self,new_value):
+        if self.is_switching_mode == False:
+            self.currentConfiguration.analog_gain = new_value
+            self.configurationManager.update_configuration(self.currentConfiguration.id,'AnalogGain',new_value)
+            self.signal_newAnalogGain.emit(new_value)
+
+    def update_config_illumination_intensity(self,new_value):
+        if self.is_switching_mode == False:
+            self.currentConfiguration.illumination_intensity = new_value
+            self.configurationManager.update_configuration(self.currentConfiguration.id,'IlluminationIntensity',new_value)
+            self.liveController.set_illumination(self.currentConfiguration.illumination_source, self.currentConfiguration.illumination_intensity)
+
+    def set_microscope_mode(self,config):
+        # self.liveController.set_microscope_mode(config)
+        self.dropdown_modeSelection.setCurrentText(config.name)
+
+    def set_trigger_mode(self,trigger_mode):
+        self.dropdown_triggerManu.setCurrentText(trigger_mode)
+        self.liveController.set_trigger_mode(self.dropdown_triggerManu.currentText())
