+++ software/control/widgets/navigation.py
@@ -1,0 +1,256 @@
+# qt libraries
+from qtpy.QtCore import *
+from qtpy.QtWidgets import *
+from qtpy.QtGui import *
+
+import pyqtgraph as pg
+
+from datetime import datetime
+
+from control._def import *
+import control.core as core
+
+from typing import Optional, Union, List, Tuple
+
+class NavigationWidget(QFrame):
+    def __init__(self, navigationController, slidePositionController=None, main=None, widget_configuration = 'full', *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.navigationController = navigationController
+        self.slidePositionController = slidePositionController
+        self.widget_configuration = widget_configuration
+        self.slide_position = None
+        self.add_components()
+        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
+
+    def add_components(self):
+        self.label_Xpos = QLabel()
+        self.label_Xpos.setNum(0)
+        self.label_Xpos.setFrameStyle(QFrame.Panel | QFrame.Sunken)
+        self.entry_dX = QDoubleSpinBox()
+        self.entry_dX.setMinimum(0) 
+        self.entry_dX.setMaximum(25) 
+        self.entry_dX.setSingleStep(0.2)
+        self.entry_dX.setValue(0)
+        self.entry_dX.setDecimals(3)
+        self.entry_dX.setKeyboardTracking(False)
+        self.btn_moveX_forward = QPushButton('Forward')
+        self.btn_moveX_forward.setDefault(False)
+        self.btn_moveX_backward = QPushButton('Backward')
+        self.btn_moveX_backward.setDefault(False)
+
+        self.btn_home_X = QPushButton('Home X')
+        self.btn_home_X.setDefault(False)
+        self.btn_home_X.setEnabled(HOMING_ENABLED_X)
+        self.btn_zero_X = QPushButton('Zero X')
+        self.btn_zero_X.setDefault(False)
+        
+        self.label_Ypos = QLabel()
+        self.label_Ypos.setNum(0)
+        self.label_Ypos.setFrameStyle(QFrame.Panel | QFrame.Sunken)
+        self.entry_dY = QDoubleSpinBox()
+        self.entry_dY.setMinimum(0)
+        self.entry_dY.setMaximum(25)
+        self.entry_dY.setSingleStep(0.2)
+        self.entry_dY.setValue(0)
+        self.entry_dY.setDecimals(3)
+        self.entry_dY.setKeyboardTracking(False)
+        self.btn_moveY_forward = QPushButton('Forward')
+        self.btn_moveY_forward.setDefault(False)
+        self.btn_moveY_backward = QPushButton('Backward')
+        self.btn_moveY_backward.setDefault(False)
+
+        self.btn_home_Y = QPushButton('Home Y')
+        self.btn_home_Y.setDefault(False)
+        self.btn_home_Y.setEnabled(HOMING_ENABLED_Y)
+        self.btn_zero_Y = QPushButton('Zero Y')
+        self.btn_zero_Y.setDefault(False)
+
+        self.label_Zpos = QLabel()
+        self.label_Zpos.setNum(0)
+        self.label_Zpos.setFrameStyle(QFrame.Panel | QFrame.Sunken)
+        self.entry_dZ = QDoubleSpinBox()
+        self.entry_dZ.setMinimum(0) 
+        self.entry_dZ.setMaximum(1000) 
+        self.entry_dZ.setSingleStep(0.2)
+        self.entry_dZ.setValue(0)
+        self.entry_dZ.setDecimals(3)
+        self.entry_dZ.setKeyboardTracking(False)
+        self.btn_moveZ_forward = QPushButton('Forward')
+        self.btn_moveZ_forward.setDefault(False)
+        self.btn_moveZ_backward = QPushButton('Backward')
+        self.btn_moveZ_backward.setDefault(False)
+
+        self.btn_home_Z = QPushButton('Home Z')
+        self.btn_home_Z.setDefault(False)
+        self.btn_home_Z.setEnabled(HOMING_ENABLED_Z)
+        self.btn_zero_Z = QPushButton('Zero Z')
+        self.btn_zero_Z.setDefault(False)
+
+        self.btn_load_slide = QPushButton('To Slide Loading Position')
+        
+        grid_line0 = QGridLayout()
+        grid_line0.addWidget(QLabel('X (mm)'), 0,0)
+        grid_line0.addWidget(self.label_Xpos, 0,1)
+        grid_line0.addWidget(self.entry_dX, 0,2)
+        grid_line0.addWidget(self.btn_moveX_forward, 0,3)
+        grid_line0.addWidget(self.btn_moveX_backward, 0,4)
+        
+        grid_line1 = QGridLayout()
+        grid_line1.addWidget(QLabel('Y (mm)'), 0,0)
+        grid_line1.addWidget(self.label_Ypos, 0,1)
+        grid_line1.addWidget(self.entry_dY, 0,2)
+        grid_line1.addWidget(self.btn_moveY_forward, 0,3)
+        grid_line1.addWidget(self.btn_moveY_backward, 0,4)
+
+        grid_line2 = QGridLayout()
+        grid_line2.addWidget(QLabel('Z (um)'), 0,0)
+        grid_line2.addWidget(self.label_Zpos, 0,1)
+        grid_line2.addWidget(self.entry_dZ, 0,2)
+        grid_line2.addWidget(self.btn_moveZ_forward, 0,3)
+        grid_line2.addWidget(self.btn_moveZ_backward, 0,4)
+        
+        grid_line3 = QGridLayout()
+        if self.widget_configuration == 'full':
+            grid_line3.addWidget(self.btn_zero_X, 0,3)
+            grid_line3.addWidget(self.btn_zero_Y, 0,4)
+            grid_line3.addWidget(self.btn_zero_Z, 0,5)
+            grid_line3.addWidget(self.btn_home_X, 0,0)
+            grid_line3.addWidget(self.btn_home_Y, 0,1)
+            grid_line3.addWidget(self.btn_home_Z, 0,2)
+        elif self.widget_configuration == 'malaria':
+            grid_line3.addWidget(self.btn_load_slide, 0,0,1,2)
+            grid_line3.addWidget(self.btn_home_Z, 0,2,1,1)
+            grid_line3.addWidget(self.btn_zero_Z, 0,3,1,1)
+        elif self.widget_configuration == '384 well plate':
+            grid_line3.addWidget(self.btn_home_Z, 0,2,1,1)
+            grid_line3.addWidget(self.btn_zero_Z, 0,3,1,1)
+        elif self.widget_configuration == '96 well plate':
+            grid_line3.addWidget(self.btn_home_Z, 0,2,1,1)
+            grid_line3.addWidget(self.btn_zero_Z, 0,3,1,1)
+
+        self.grid = QGridLayout()
+        self.grid.addLayout(grid_line0,0,0)
+        self.grid.addLayout(grid_line1,1,0)
+        self.grid.addLayout(grid_line2,2,0)
+        self.grid.addLayout(grid_line3,3,0)
+        self.setLayout(self.grid)
+
+        self.entry_dX.valueChanged.connect(self.set_deltaX)
+        self.entry_dY.valueChanged.connect(self.set_deltaY)
+        self.entry_dZ.valueChanged.connect(self.set_deltaZ)
+
+        self.btn_moveX_forward.clicked.connect(self.move_x_forward)
+        self.btn_moveX_backward.clicked.connect(self.move_x_backward)
+        self.btn_moveY_forward.clicked.connect(self.move_y_forward)
+        self.btn_moveY_backward.clicked.connect(self.move_y_backward)
+        self.btn_moveZ_forward.clicked.connect(self.move_z_forward)
+        self.btn_moveZ_backward.clicked.connect(self.move_z_backward)
+
+        self.btn_home_X.clicked.connect(self.home_x)
+        self.btn_home_Y.clicked.connect(self.home_y)
+        self.btn_home_Z.clicked.connect(self.home_z)
+        self.btn_zero_X.clicked.connect(self.zero_x)
+        self.btn_zero_Y.clicked.connect(self.zero_y)
+        self.btn_zero_Z.clicked.connect(self.zero_z)
+
+        self.btn_load_slide.clicked.connect(self.switch_position)
+        self.btn_load_slide.setStyleSheet("background-color: #C2C2FF");
+        
+    def move_x_forward(self):
+        self.navigationController.move_x(self.entry_dX.value())
+    def move_x_backward(self):
+        self.navigationController.move_x(-self.entry_dX.value())
+    def move_y_forward(self):
+        self.navigationController.move_y(self.entry_dY.value())
+    def move_y_backward(self):
+        self.navigationController.move_y(-self.entry_dY.value())
+    def move_z_forward(self):
+        self.navigationController.move_z(self.entry_dZ.value()/1000)
+    def move_z_backward(self):
+        self.navigationController.move_z(-self.entry_dZ.value()/1000) 
+
+    def set_deltaX(self,value):
+        mm_per_ustep = SCREW_PITCH_X_MM/(self.navigationController.x_microstepping*FULLSTEPS_PER_REV_X) # to implement a get_x_microstepping() in multipointController
+        deltaX = round(value/mm_per_ustep)*mm_per_ustep
+        self.entry_dX.setValue(deltaX)
+    def set_deltaY(self,value):
+        mm_per_ustep = SCREW_PITCH_Y_MM/(self.navigationController.y_microstepping*FULLSTEPS_PER_REV_Y)
+        deltaY = round(value/mm_per_ustep)*mm_per_ustep
+        self.entry_dY.setValue(deltaY)
+    def set_deltaZ(self,value):
+        mm_per_ustep = SCREW_PITCH_Z_MM/(self.navigationController.z_microstepping*FULLSTEPS_PER_REV_Z)
+        deltaZ = round(value/1000/mm_per_ustep)*mm_per_ustep*1000
+        self.entry_dZ.setValue(deltaZ)
+
+    def home_x(self):
+        msg = QMessageBox()
+        msg.setIcon(QMessageBox.Information)
+        msg.setText("Confirm your action")
+        msg.setInformativeText("Click OK to run homing")
+        msg.setWindowTitle("Confirmation")
+        msg.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
+        msg.setDefaultButton(QMessageBox.Cancel)
+        retval = msg.exec_()
+        if QMessageBox.Ok == retval:
+            self.navigationController.home_x()
+
+    def home_y(self):
+        msg = QMessageBox()
+        msg.setIcon(QMessageBox.Information)
+        msg.setText("Confirm your action")
+        msg.setInformativeText("Click OK to run homing")
+        msg.setWindowTitle("Confirmation")
+        msg.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
+        msg.setDefaultButton(QMessageBox.Cancel)
+        retval = msg.exec_()
+        if QMessageBox.Ok == retval:
+            self.navigationController.home_y()
+
+    def home_z(self):
+        msg = QMessageBox()
+        msg.setIcon(QMessageBox.Information)
+        msg.setText("Confirm your action")
+        msg.setInformativeText("Click OK to run homing")
+        msg.setWindowTitle("Confirmation")
+        msg.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
+        msg.setDefaultButton(QMessageBox.Cancel)
+        retval = msg.exec_()
+        if QMessageBox.Ok == retval:
+            self.navigationController.home_z()
+
+    def zero_x(self):
+        self.navigationController.zero_x()
+
+    def zero_y(self):
+        self.navigationController.zero_y()
+
+    def zero_z(self):
+        self.navigationController.zero_z()
+
+    def slot_slide_loading_position_reached(self):
+        self.slide_position = 'loading'
+        self.btn_load_slide.setStyleSheet("background-color: #C2FFC2");
+        self.btn_load_slide.setText('To Slide Scanning Position')
+        self.btn_moveX_forward.setEnabled(False)
+        self.btn_moveX_backward.setEnabled(False)
+        self.btn_moveY_forward.setEnabled(False)
+        self.btn_moveY_backward.setEnabled(False)
+        self.btn_moveZ_forward.setEnabled(False)
+        self.btn_moveZ_backward.setEnabled(False)
+
+    def slot_slide_scanning_position_reached(self):
+        self.slide_position = 'scanning'
+        self.btn_load_slide.setStyleSheet("background-color: #C2C2FF");
+        self.btn_load_slide.setText('To Slide Loading Position')
+        self.btn_moveX_forward.setEnabled(True)
+        self.btn_moveX_backward.setEnabled(True)
+        self.btn_moveY_forward.setEnabled(True)
+        self.btn_moveY_backward.setEnabled(True)
+        self.btn_moveZ_forward.setEnabled(True)
+        self.btn_moveZ_backward.setEnabled(True)
+
+    def switch_position(self):
+        if self.slide_position != 'loading':
+            self.slidePositionController.move_to_slide_loading_position()
+        else:
+            self.slidePositionController.move_to_slide_scanning_position()
