+++ software/control/widgets/recording.py
@@ -1,0 +1,117 @@
+# qt libraries
+from qtpy.QtCore import *
+from qtpy.QtWidgets import *
+from qtpy.QtGui import *
+
+import pyqtgraph as pg
+
+from datetime import datetime
+
+from control._def import *
+import control.core as core
+
+from typing import Optional, Union, List, Tuple
+
+class RecordingWidget(QFrame):
+    def __init__(self, streamHandler, imageSaver, main=None, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.imageSaver = imageSaver # for saving path control
+        self.streamHandler = streamHandler
+        self.base_path_is_set = False
+        self.add_components()
+        self.setFrameStyle(QFrame.Panel | QFrame.Raised)
+
+    def add_components(self):
+        self.btn_setSavingDir = QPushButton('Browse')
+        self.btn_setSavingDir.setDefault(False)
+        self.btn_setSavingDir.setIcon(QIcon('icon/folder.png'))
+        
+        self.lineEdit_savingDir = QLineEdit()
+        self.lineEdit_savingDir.setReadOnly(True)
+        self.lineEdit_savingDir.setText('Choose a base saving directory')
+
+        self.lineEdit_savingDir.setText(DEFAULT_SAVING_PATH)
+        self.imageSaver.set_base_path(DEFAULT_SAVING_PATH)
+
+        self.lineEdit_experimentID = QLineEdit()
+
+        self.entry_saveFPS = QDoubleSpinBox()
+        self.entry_saveFPS.setMinimum(0.02) 
+        self.entry_saveFPS.setMaximum(1000) 
+        self.entry_saveFPS.setSingleStep(1)
+        self.entry_saveFPS.setValue(1)
+        self.streamHandler.set_save_fps(1)
+
+        self.entry_timeLimit = QSpinBox()
+        self.entry_timeLimit.setMinimum(-1) 
+        self.entry_timeLimit.setMaximum(60*60*24*30) 
+        self.entry_timeLimit.setSingleStep(1)
+        self.entry_timeLimit.setValue(-1)
+
+        self.btn_record = QPushButton("Record")
+        self.btn_record.setCheckable(True)
+        self.btn_record.setChecked(False)
+        self.btn_record.setDefault(False)
+
+        grid_line1 = QGridLayout()
+        grid_line1.addWidget(QLabel('Saving Path'))
+        grid_line1.addWidget(self.lineEdit_savingDir, 0,1)
+        grid_line1.addWidget(self.btn_setSavingDir, 0,2)
+
+        grid_line2 = QGridLayout()
+        grid_line2.addWidget(QLabel('Experiment ID'), 0,0)
+        grid_line2.addWidget(self.lineEdit_experimentID,0,1)
+
+        grid_line3 = QGridLayout()
+        grid_line3.addWidget(QLabel('Saving FPS'), 0,0)
+        grid_line3.addWidget(self.entry_saveFPS, 0,1)
+        grid_line3.addWidget(QLabel('Time Limit (s)'), 0,2)
+        grid_line3.addWidget(self.entry_timeLimit, 0,3)
+        grid_line3.addWidget(self.btn_record, 0,4)
+
+        self.grid = QGridLayout()
+        self.grid.addLayout(grid_line1,0,0)
+        self.grid.addLayout(grid_line2,1,0)
+        self.grid.addLayout(grid_line3,2,0)
+        self.setLayout(self.grid)
+
+        # add and display a timer - to be implemented
+        # self.timer = QTimer()
+
+        # connections
+        self.btn_setSavingDir.clicked.connect(self.set_saving_dir)
+        self.btn_record.clicked.connect(self.toggle_recording)
+        self.entry_saveFPS.valueChanged.connect(self.streamHandler.set_save_fps)
+        self.entry_timeLimit.valueChanged.connect(self.imageSaver.set_recording_time_limit)
+        self.imageSaver.stop_recording.connect(self.stop_recording)
+
+    def set_saving_dir(self):
+        dialog = QFileDialog()
+        save_dir_base = dialog.getExistingDirectory(None, "Select Folder")
+        self.imageSaver.set_base_path(save_dir_base)
+        self.lineEdit_savingDir.setText(save_dir_base)
+        self.base_path_is_set = True
+
+    def toggle_recording(self,pressed):
+        if self.base_path_is_set == False:
+            self.btn_record.setChecked(False)
+            msg = QMessageBox()
+            msg.setText("Please choose base saving directory first")
+            msg.exec_()
+            return
+        if pressed:
+            self.lineEdit_experimentID.setEnabled(False)
+            self.btn_setSavingDir.setEnabled(False)
+            self.imageSaver.start_new_experiment(self.lineEdit_experimentID.text())
+            self.streamHandler.start_recording()
+        else:
+            self.streamHandler.stop_recording()
+            self.lineEdit_experimentID.setEnabled(True)
+            self.btn_setSavingDir.setEnabled(True)
+
+    # stop_recording can be called by imageSaver
+    def stop_recording(self):
+        self.lineEdit_experimentID.setEnabled(True)
+        self.btn_record.setChecked(False)
+        self.streamHandler.stop_recording()
+        self.btn_setSavingDir.setEnabled(True)
